@@ -1,94 +1,92 @@
 name: revanced_manager
-description: The official ReVanced Manager.
-homepage: https://github.com/revanced/revanced-manager
+description: "Application to use ReVanced on Android."
+homepage: https://revanced.app
 
 publish_to: 'none'
 
-version: 1.1.0+100100000
+version: 1.23.2+101800050
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: ^3.5.3
 
 dependencies:
-  animations: ^2.0.7
-  app_installer: ^1.1.0
-  collection: ^1.17.0
-  cross_connectivity: ^3.0.5
-  cr_file_saver:
-    git:
-      url: https://github.com/dhruvanbhalara/cr_file_saver.git
-      ref: a08326ecb48f581b4b09e2e2665d31ed1704c7af
+  animations: ^2.0.11
+  collection: ^1.18.0
+  connectivity_plus: ^6.0.3
   device_apps:
     git:
       url: https://github.com/ponces/flutter_plugin_device_apps
-      ref: revanced-manager
-  device_info_plus: ^8.1.0
-  dynamic_color: ^1.6.3
-  dio: ^5.0.0
+      ref: 0609662324b9e00931fccfa87e9e1b00b7610907
+  device_info_plus: ^10.1.0
+  dio: ^5.4.3+1
+  dio_cache_interceptor: ^3.5.0
+  dynamic_color: ^1.7.0
   dynamic_themes: ^1.1.0
   expandable: ^5.0.1
-  file_picker:
-    git:
-      url: https://github.com/alexmercerind/flutter_file_picker
-      ref: master
-  flex_color_scheme: ^7.0.1
+  file_picker: ^8.0.5
   flutter:
     sdk: flutter
-  flutter_background: ^1.2.0
-  flutter_cache_manager: ^3.3.0
-  flutter_i18n: ^0.32.4
-  flutter_local_notifications: ^13.0.0
+  flutter_background: ^1.3.0+1
+  flutter_cache_manager: ^3.3.2
+  flutter_file_dialog: ^3.0.2
+  flutter_local_notifications: ^17.1.0
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^2.0.4
-  fluttertoast: ^8.2.1
-  font_awesome_flutter: ^10.4.0
-  get_it: 7.2.0
-  google_fonts: ^4.0.3
-  http: ^0.13.5
-  injectable: ^2.1.1
-  intl: ^0.17.0
-  json_annotation: ^4.8.0
+  flutter_markdown: ^0.7.3
+  fluttertoast: ^8.2.5
+  font_awesome_flutter: ^10.7.0
+  google_fonts: ^6.2.1
+  injectable: ^2.4.0
+  intl: 0.19.0
+  json_annotation: ^4.9.0
+  language_code: ^0.5.3+2
   logcat:
     git:
-      url: https://github.com/SuaMusica/logcat
-      ref: feature/nullSafe
-  package_info_plus: ^3.0.3
-  path_provider: ^2.0.14
-  permission_handler: ^10.2.0
-  pull_to_refresh: ^2.0.0
+      url: https://github.com/BenjaminHalko/logcat
+      ref: 4a6d5e0e22292c8eb160cfb9365b9ea29735fd43 # Branch: master
+  package_info_plus: ^8.0.2
+  path_provider: ^2.1.3
+  permission_handler: ^11.3.1
   root:
     git:
-      url: https://github.com/gokul1630/root
-      ref: main
-  share_extend: ^2.0.0
-  shared_preferences: ^2.1.0
-  skeletons: ^0.0.3
-  stacked: ^3.2.0
-  stacked_generator: ^1.1.0
-  stacked_services: ^1.0.0
-  stacked_themes: ^0.3.10
-  timeago: ^3.3.0
-  timezone: ^0.9.0
-  url_launcher: ^6.1.10
-  wakelock: ^0.6.2
-  flutter_dotenv: ^5.0.2
-  flutter_markdown: ^0.6.14
-  pub_release: ^8.0.3
-  dio_cache_interceptor: ^3.4.0
+      url: https://github.com/validcube/root
+      ref: 68e5678a535a2a3344828a14a39017fa74b9098c # Branch: libsu-521
+  screenshot_callback:
+    git: # remove once https://github.com/flutter-moum/flutter_screenshot_callback/pull/81 is merged
+      url: https://github.com/BenjaminHalko/flutter_screenshot_callback
+      ref: 1a1616ac91e16cd1f3dd170a81febf27ffce3587 # Branch: master
+  share_plus: ^10.0.2
+  shared_preferences: ^2.2.3
+  skeletons:
+    git: # remove once https://github.com/badjio/skeletons/pull/11 is merged
+      url: https://github.com/Ofceab-Studio/skeletons
+      ref: 326fbb4223ac4d8a6301cc2f16b6112ecd308c71 # Branch: master
+  slang: ^3.30.2
+  slang_flutter: ^3.30.0
+  stacked: ^3.4.2
+  stacked_generator: ^1.6.0
+  stacked_services: ^1.5.0
+  synchronized: ^3.1.0+1
+  timeago: ^3.6.1
+  timezone: ^0.9.3
+  url_launcher: ^6.2.6
+  wakelock_plus: ^1.2.4
 
 dev_dependencies:
-  json_serializable: ^6.6.1
-  build_runner: any
-  flutter_launcher_icons: ^0.13.0
-  flutter_lints: ^2.0.1
-  flutter_test:
-    sdk: flutter
-  injectable_generator: ^2.1.5  
-
-
+  analyzer: ^6.4.1
+  build_runner: ^2.4.12
+  flutter_lints: ^5.0.0
+  injectable_generator: ^2.6.1
+  json_serializable: ^6.8.0
 
 flutter:
   uses-material-design: true
+  fonts:
+    - family: CustomIcon
+      fonts:
+        - asset: fonts/custom-icons.ttf
   assets:
     - assets/i18n/
+
+flutter_localizations:
+  sdk: flutter
