@@ -1,67 +1,52 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
+import 'package:revanced_manager/services/download_manager.dart';
 import 'package:revanced_manager/services/github_api.dart';
 import 'package:revanced_manager/services/manager_api.dart';
-import 'package:revanced_manager/services/patcher_api.dart';
 import 'package:revanced_manager/services/revanced_api.dart';
+import 'package:revanced_manager/services/root_api.dart';
 import 'package:revanced_manager/ui/theme/dynamic_theme_builder.dart';
 import 'package:revanced_manager/ui/views/navigation/navigation_view.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:stacked_themes/stacked_themes.dart';
 import 'package:timezone/data/latest.dart' as tz;
 
 late SharedPreferences prefs;
 Future main() async {
-  await ThemeManager.initialise();
   await setupLocator();
   WidgetsFlutterBinding.ensureInitialized();
   await locator<ManagerAPI>().initialize();
+
+  await locator<DownloadManager>().initialize();
   final String apiUrl = locator<ManagerAPI>().getApiUrl();
   await locator<RevancedAPI>().initialize(apiUrl);
   final String repoUrl = locator<ManagerAPI>().getRepoUrl();
   locator<GithubAPI>().initialize(repoUrl);
-  await locator<PatcherAPI>().initialize();
   tz.initializeTimeZones();
+
+  // TODO(aAbed): remove in the future, keep it for now during migration.
+  final rootAPI = RootAPI();
+  if (await rootAPI.hasRootPermissions()) {
+    await rootAPI.removeOrphanedFiles();
+  }
+
   prefs = await SharedPreferences.getInstance();
 
-  runApp(const MyApp());
+  final managerAPI = locator<ManagerAPI>();
+  final locale = managerAPI.getLocale();
+  LocaleSettings.setLocaleRaw(locale);
+
+  runApp(TranslationProvider(child: const MyApp()));
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    // String rawLocale = prefs.getString('language') ?? 'en_US';
-    // String replaceLocale = rawLocale.replaceAll('_', '-');
-    // List<String> localeList = replaceLocale.split('-');
-    // Locale locale = Locale(localeList[0], localeList[1]);
-    const Locale locale = Locale('en', 'US');
-
-    return DynamicThemeBuilder(
+    return const DynamicThemeBuilder(
       title: 'ReVanced Manager',
-      home: const NavigationView(),
-      localizationsDelegates: [
-        FlutterI18nDelegate(
-          translationLoader: FileTranslationLoader(
-            fallbackFile: 'en_US',
-            forcedLocale: locale,
-            basePath: 'assets/i18n',
-            useCountryCode: true,
-          ),
-          missingTranslationHandler: (key, locale) {
-            log(
-              '--> Missing translation: key: $key, languageCode: ${locale?.languageCode}',
-            );
-          },
-        ),
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate
-      ],
+      home: NavigationView(),
     );
   }
 }
