@@ -2,15 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 var lightCustomColorScheme = ColorScheme.fromSeed(
-  seedColor: Colors.blue,
-  primary: const Color(0xff1B73E8),
+  seedColor: Colors.purple,
+  primary: const Color(0xFF4C51C0),
 );
 
 var lightCustomTheme = ThemeData(
   useMaterial3: true,
   colorScheme: lightCustomColorScheme,
   navigationBarTheme: NavigationBarThemeData(
-    labelTextStyle: MaterialStateProperty.all(
+    labelTextStyle: WidgetStateProperty.all(
       TextStyle(
         color: lightCustomColorScheme.onSurface,
         fontWeight: FontWeight.w500,
@@ -21,24 +21,24 @@ var lightCustomTheme = ThemeData(
 );
 
 var darkCustomColorScheme = ColorScheme.fromSeed(
-  seedColor: Colors.blue,
+  seedColor: Colors.purple,
   brightness: Brightness.dark,
-  primary: const Color(0xffA5CAFF),
-  surface: const Color(0xff1B1A1D),
+  primary: const Color(0xFFBFC1FF),
+  surface: const Color(0xFF131316),
 );
 
 var darkCustomTheme = ThemeData(
   useMaterial3: true,
   colorScheme: darkCustomColorScheme,
   navigationBarTheme: NavigationBarThemeData(
-    labelTextStyle: MaterialStateProperty.all(
+    labelTextStyle: WidgetStateProperty.all(
       TextStyle(
         color: darkCustomColorScheme.onSurface,
         fontWeight: FontWeight.w500,
       ),
     ),
   ),
-  canvasColor: const Color(0xff1B1A1D),
-  scaffoldBackgroundColor: const Color(0xff1B1A1D),
+  canvasColor: const Color(0xFF131316),
+  scaffoldBackgroundColor: const Color(0xFF131316),
   textTheme: GoogleFonts.robotoTextTheme(ThemeData.dark().textTheme),
 );
