@@ -1,14 +1,15 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:flutter/material.dart' hide SearchBar;
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/views/app_selector/app_selector_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/appSelectorView/app_skeleton_loader.dart';
 import 'package:revanced_manager/ui/widgets/appSelectorView/installed_app_item.dart';
 import 'package:revanced_manager/ui/widgets/appSelectorView/not_installed_app_item.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_floating_action_button_extended.dart';
 import 'package:revanced_manager/ui/widgets/shared/search_bar.dart';
 import 'package:stacked/stacked.dart' hide SkeletonLoader;
 
 class AppSelectorView extends StatefulWidget {
-  const AppSelectorView({Key? key}) : super(key: key);
+  const AppSelectorView({super.key});
 
   @override
   State<AppSelectorView> createState() => _AppSelectorViewState();
@@ -23,13 +24,11 @@ class _AppSelectorViewState extends State<AppSelectorView> {
       onViewModelReady: (model) => model.initialize(),
       viewModelBuilder: () => AppSelectorViewModel(),
       builder: (context, model, child) => Scaffold(
-        resizeToAvoidBottomInset: false,
-        floatingActionButton: FloatingActionButton.extended(
-          label: I18nText('appSelectorView.storageButton'),
+        floatingActionButton: HapticFloatingActionButtonExtended(
+          label: Text(t.appSelectorView.storageButton),
           icon: const Icon(Icons.sd_storage),
           onPressed: () {
             model.selectAppFromStorage(context);
-            Navigator.of(context).pop();
           },
         ),
         body: CustomScrollView(
@@ -37,14 +36,12 @@ class _AppSelectorViewState extends State<AppSelectorView> {
             SliverAppBar(
               pinned: true,
               floating: true,
-              title: I18nText(
-                'appSelectorView.viewTitle',
-                child: Text(
-                  '',
-                  style: TextStyle(
-                    color: Theme.of(context).textTheme.titleLarge!.color,
-                  ),
-                ),
+              title: Text(
+                t.appSelectorView.viewTitle,
+              ),
+              titleTextStyle: TextStyle(
+                fontSize: 22.0,
+                color: Theme.of(context).textTheme.titleLarge!.color,
               ),
               leading: IconButton(
                 icon: Icon(
@@ -54,17 +51,14 @@ class _AppSelectorViewState extends State<AppSelectorView> {
                 onPressed: () => Navigator.of(context).pop(),
               ),
               bottom: PreferredSize(
-                preferredSize: const Size.fromHeight(64.0),
+                preferredSize: const Size.fromHeight(66.0),
                 child: Padding(
                   padding: const EdgeInsets.symmetric(
                     vertical: 8.0,
                     horizontal: 12.0,
                   ),
                   child: SearchBar(
-                    hintText: FlutterI18n.translate(
-                      context,
-                      'appSelectorView.searchBarHint',
-                    ),
+                    hintText: t.appSelectorView.searchBarHint,
                     onQueryChanged: (searchQuery) {
                       setState(() {
                         _query = searchQuery;
@@ -77,27 +71,24 @@ class _AppSelectorViewState extends State<AppSelectorView> {
             SliverToBoxAdapter(
               child: model.noApps
                   ? Center(
-                      child: I18nText(
-                        'appSelectorView.noAppsLabel',
-                        child: Text(
-                          '',
-                          style: TextStyle(
-                            color:
-                                Theme.of(context).textTheme.titleLarge!.color,
-                          ),
+                      child: Text(
+                        t.appSelectorCard.noAppsLabel,
+                        style: TextStyle(
+                          color: Theme.of(context).textTheme.titleLarge!.color,
                         ),
                       ),
                     )
-                  : model.apps.isEmpty
+                  : model.allApps.isEmpty && model.apps.isEmpty
                       ? const AppSkeletonLoader()
                       : Padding(
                           padding: const EdgeInsets.symmetric(horizontal: 12.0)
-                              .copyWith(bottom: 80),
+                              .copyWith(
+                            bottom:
+                                MediaQuery.viewPaddingOf(context).bottom + 8.0,
+                          ),
                           child: Column(
                             children: [
-                              ...model
-                                  .getFilteredApps(_query)
-                                  .map(
+                              ...model.getFilteredApps(_query).map(
                                     (app) => InstalledAppItem(
                                       name: app.appName,
                                       pkgName: app.packageName,
@@ -108,22 +99,18 @@ class _AppSelectorViewState extends State<AppSelectorView> {
                                           model.getSuggestedVersion(
                                         app.packageName,
                                       ),
-                                      onTap: () {
-                                        model.isRooted
-                                            ? model.selectApp(app).then(
-                                                  (_) => Navigator.of(context)
-                                                      .pop(),
-                                                )
-                                            : model.showSelectFromStorageDialog(
-                                                context,
-                                              );
-                                      },
+                                      installedVersion: app.versionName!,
+                                      onTap: () => model.canSelectInstalled(
+                                        context,
+                                        app.packageName,
+                                      ),
+                                      onLinkTap: () =>
+                                          model.searchSuggestedVersionOnWeb(
+                                        packageName: app.packageName,
+                                      ),
                                     ),
-                                  )
-                                  .toList(),
-                              ...model
-                                  .getFilteredAppsNames(_query)
-                                  .map(
+                                  ),
+                              ...model.getFilteredAppsNames(_query).map(
                                     (app) => NotInstalledAppItem(
                                       name: app,
                                       patchesCount: model.patchesCount(app),
@@ -132,9 +119,13 @@ class _AppSelectorViewState extends State<AppSelectorView> {
                                       onTap: () {
                                         model.showDownloadToast();
                                       },
+                                      onLinkTap: () =>
+                                          model.searchSuggestedVersionOnWeb(
+                                        packageName: app,
+                                      ),
                                     ),
-                                  )
-                                  .toList(),
+                                  ),
+                              const SizedBox(height: 70.0),
                             ],
                           ),
                         ),
