@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/views/contributors/contributors_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/contributorsView/contributors_card.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_sliver_app_bar.dart';
 import 'package:stacked/stacked.dart';
 
 class ContributorsView extends StatelessWidget {
-  const ContributorsView({Key? key}) : super(key: key);
+  const ContributorsView({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -18,13 +18,10 @@ class ContributorsView extends StatelessWidget {
         body: CustomScrollView(
           slivers: <Widget>[
             CustomSliverAppBar(
-              title: I18nText(
-                'contributorsView.widgetTitle',
-                child: Text(
-                  '',
-                  style: GoogleFonts.inter(
-                    color: Theme.of(context).textTheme.titleLarge!.color,
-                  ),
+              title: Text(
+                t.contributorsView.widgetTitle,
+                style: GoogleFonts.inter(
+                  color: Theme.of(context).textTheme.titleLarge!.color,
                 ),
               ),
             ),
@@ -33,30 +30,14 @@ class ContributorsView extends StatelessWidget {
               sliver: SliverList(
                 delegate: SliverChildListDelegate.fixed(
                   <Widget>[
-                    ContributorsCard(
-                      title: 'contributorsView.patcherContributors',
-                      contributors: model.patcherContributors,
-                    ),
-                    const SizedBox(height: 20),
-                    ContributorsCard(
-                      title: 'contributorsView.patchesContributors',
-                      contributors: model.patchesContributors,
-                    ),
-                    const SizedBox(height: 20),
-                    ContributorsCard(
-                      title: 'contributorsView.integrationsContributors',
-                      contributors: model.integrationsContributors,
-                    ),
-                    const SizedBox(height: 20),
-                    ContributorsCard(
-                      title: 'contributorsView.cliContributors',
-                      contributors: model.cliContributors,
-                    ),
-                    const SizedBox(height: 20),
-                    ContributorsCard(
-                      title: 'contributorsView.managerContributors',
-                      contributors: model.managerContributors,
-                    ),
+                    for (final String tool in model.contributors.keys) ...[
+                      ContributorsCard(
+                        title: tool,
+                        contributors: model.contributors[tool]!,
+                      ),
+                      const SizedBox(height: 20),
+                    ],
+                    SizedBox(height: MediaQuery.viewPaddingOf(context).bottom),
                   ],
                 ),
               ),
