@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/views/patcher/patcher_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/patcherView/app_selector_card.dart';
 import 'package:revanced_manager/ui/widgets/patcherView/patch_selector_card.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_sliver_app_bar.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_floating_action_button_extended.dart';
 import 'package:stacked/stacked.dart';
 
 class PatcherView extends StatelessWidget {
-  const PatcherView({Key? key}) : super(key: key);
+  const PatcherView({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -19,23 +20,27 @@ class PatcherView extends StatelessWidget {
       builder: (context, model, child) => Scaffold(
         floatingActionButton: Visibility(
           visible: model.showPatchButton(),
-          child: FloatingActionButton.extended(
-            label: I18nText('patcherView.patchButton'),
+          child: HapticFloatingActionButtonExtended(
+            label: Text(t.patcherView.patchButton),
             icon: const Icon(Icons.build),
-            onPressed: () => model.showPatchConfirmationDialog(context),
+            onPressed: () async {
+              if (model.checkRequiredPatchOption(context)) {
+                final bool proceed = model.showRemovedPatchesDialog(context);
+                if (proceed && context.mounted) {
+                  model.showIncompatibleArchWarningDialog(context);
+                }
+              }
+            },
           ),
         ),
         body: CustomScrollView(
           slivers: <Widget>[
             CustomSliverAppBar(
               isMainView: true,
-              title: I18nText(
-                'patcherView.widgetTitle',
-                child: Text(
-                  '',
-                  style: GoogleFonts.inter(
-                    color: Theme.of(context).textTheme.titleLarge!.color,
-                  ),
+              title: Text(
+                t.patcherView.widgetTitle,
+                style: GoogleFonts.inter(
+                  color: Theme.of(context).textTheme.titleLarge!.color,
                 ),
               ),
             ),
@@ -45,7 +50,10 @@ class PatcherView extends StatelessWidget {
                 delegate: SliverChildListDelegate.fixed(
                   <Widget>[
                     AppSelectorCard(
-                      onPressed: () => model.navigateToAppSelector(),
+                      onPressed: () => {
+                        model.navigateToAppSelector(),
+                        model.ctx = context,
+                      },
                     ),
                     const SizedBox(height: 16),
                     Opacity(
