@@ -1,14 +1,14 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:flutter/material.dart' hide SearchBar;
+import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
+import 'package:revanced_manager/services/manager_api.dart';
 import 'package:revanced_manager/ui/views/patches_selector/patches_selector_viewmodel.dart';
-import 'package:revanced_manager/ui/widgets/patchesSelectorView/patch_item.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_popup_menu.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_floating_action_button_extended.dart';
 import 'package:revanced_manager/ui/widgets/shared/search_bar.dart';
-import 'package:revanced_manager/utils/check_for_supported_patch.dart';
 import 'package:stacked/stacked.dart';
 
 class PatchesSelectorView extends StatefulWidget {
-  const PatchesSelectorView({Key? key}) : super(key: key);
+  const PatchesSelectorView({super.key});
 
   @override
   State<PatchesSelectorView> createState() => _PatchesSelectorViewState();
@@ -16,6 +16,18 @@ class PatchesSelectorView extends StatefulWidget {
 
 class _PatchesSelectorViewState extends State<PatchesSelectorView> {
   String _query = '';
+  final _managerAPI = locator<ManagerAPI>();
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      if (!_managerAPI.isPatchesChangeEnabled() &&
+          _managerAPI.showPatchesChangeWarning()) {
+        _managerAPI.showPatchesChangeWarningDialog(context);
+      }
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -23,20 +35,21 @@ class _PatchesSelectorViewState extends State<PatchesSelectorView> {
       onViewModelReady: (model) => model.initialize(),
       viewModelBuilder: () => PatchesSelectorViewModel(),
       builder: (context, model, child) => Scaffold(
-        resizeToAvoidBottomInset: false,
         floatingActionButton: Visibility(
           visible: model.patches.isNotEmpty,
-          child: FloatingActionButton.extended(
+          child: HapticFloatingActionButtonExtended(
             label: Row(
               children: <Widget>[
-                I18nText('patchesSelectorView.doneButton'),
-                Text(' (${model.selectedPatches.length})')
+                Text(t.patchesSelectorView.doneButton),
+                Text(' (${model.selectedPatches.length})'),
               ],
             ),
             icon: const Icon(Icons.check),
             onPressed: () {
-              model.selectPatches();
-              Navigator.of(context).pop();
+              if (!model.areRequiredOptionsNull(context)) {
+                model.selectPatches();
+                Navigator.of(context).pop();
+              }
             },
           ),
         ),
@@ -45,32 +58,33 @@ class _PatchesSelectorViewState extends State<PatchesSelectorView> {
             SliverAppBar(
               pinned: true,
               floating: true,
-              title: I18nText(
-                'patchesSelectorView.viewTitle',
-                child: Text(
-                  '',
-                  style: TextStyle(
-                    color: Theme.of(context).textTheme.titleLarge!.color,
-                  ),
-                ),
+              title: Text(
+                t.patchesSelectorView.viewTitle,
+              ),
+              titleTextStyle: TextStyle(
+                fontSize: 22.0,
+                color: Theme.of(context).textTheme.titleLarge!.color,
               ),
               leading: IconButton(
                 icon: Icon(
                   Icons.arrow_back,
                   color: Theme.of(context).textTheme.titleLarge!.color,
                 ),
-                onPressed: () => Navigator.of(context).pop(),
+                onPressed: () {
+                  model.resetSelection();
+                  Navigator.of(context).pop();
+                },
               ),
               actions: [
                 Container(
-                  margin: const EdgeInsets.only(top: 12, bottom: 12),
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 6, vertical: 6),
+                  margin: const EdgeInsets.symmetric(vertical: 12),
+                  padding: const EdgeInsets.symmetric(horizontal: 6),
                   decoration: BoxDecoration(
                     color:
                         Theme.of(context).colorScheme.tertiary.withOpacity(0.5),
                     borderRadius: BorderRadius.circular(6),
                   ),
+                  alignment: Alignment.center,
                   child: Text(
                     model.patchesVersion!,
                     style: TextStyle(
@@ -78,33 +92,29 @@ class _PatchesSelectorViewState extends State<PatchesSelectorView> {
                     ),
                   ),
                 ),
-                CustomPopupMenu(
-                  onSelected: (value) => {model.onMenuSelection(value)},
-                  children: {
-                    0: I18nText(
-                      'patchesSelectorView.loadPatchesSelection',
-                      child: const Text(
-                        '',
-                        style: TextStyle(
-                          fontWeight: FontWeight.bold,
-                        ),
+                PopupMenuButton(
+                  onSelected: (value) {
+                    model.onMenuSelection(value, context);
+                  },
+                  itemBuilder: (BuildContext context) => <PopupMenuEntry>[
+                    PopupMenuItem(
+                      value: 0,
+                      child: Text(
+                        t.patchesSelectorView.loadPatchesSelection,
                       ),
                     ),
-                  },
+                  ],
                 ),
               ],
               bottom: PreferredSize(
-                preferredSize: const Size.fromHeight(64.0),
+                preferredSize: const Size.fromHeight(66.0),
                 child: Padding(
                   padding: const EdgeInsets.symmetric(
                     vertical: 8.0,
                     horizontal: 12.0,
                   ),
                   child: SearchBar(
-                    hintText: FlutterI18n.translate(
-                      context,
-                      'patchesSelectorView.searchBarHint',
-                    ),
+                    hintText: t.patchesSelectorView.searchBarHint,
                     onQueryChanged: (searchQuery) {
                       setState(() {
                         _query = searchQuery;
@@ -119,63 +129,105 @@ class _PatchesSelectorViewState extends State<PatchesSelectorView> {
                   ? Padding(
                       padding: const EdgeInsets.all(8.0),
                       child: Center(
-                        child: I18nText(
-                          'patchesSelectorView.noPatchesFound',
-                          child: Text(
-                            '',
-                            style: Theme.of(context).textTheme.bodyMedium,
-                          ),
+                        child: Text(
+                          t.patchesSelectorView.noPatchesFound,
+                          style: Theme.of(context).textTheme.bodyMedium,
                         ),
                       ),
                     )
                   : Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 12.0)
-                          .copyWith(bottom: 80),
+                      padding:
+                          const EdgeInsets.symmetric(horizontal: 12.0).copyWith(
+                        bottom: MediaQuery.viewPaddingOf(context).bottom + 8.0,
+                      ),
                       child: Column(
                         children: [
                           Row(
                             children: [
                               ActionChip(
-                                label: I18nText('patchesSelectorView.default'),
-                                tooltip: FlutterI18n.translate(
-                                  context,
-                                  'patchesSelectorView.defaultTooltip',
-                                ),
+                                label: Text(t.patchesSelectorView.defaultChip),
+                                tooltip: t.patchesSelectorView.defaultTooltip,
                                 onPressed: () {
-                                  model.selectDefaultPatches();
+                                  if (_managerAPI.isPatchesChangeEnabled()) {
+                                    model.selectDefaultPatches();
+                                  } else {
+                                    model.showPatchesChangeDialog(context);
+                                  }
                                 },
                               ),
                               const SizedBox(width: 8),
                               ActionChip(
-                                label: I18nText('patchesSelectorView.none'),
-                                tooltip: FlutterI18n.translate(
-                                  context,
-                                  'patchesSelectorView.noneTooltip',
-                                ),
+                                label: Text(t.patchesSelectorView.noneChip),
+                                tooltip: t.patchesSelectorView.noneTooltip,
                                 onPressed: () {
-                                  model.clearPatches();
+                                  if (_managerAPI.isPatchesChangeEnabled()) {
+                                    model.clearPatches();
+                                  } else {
+                                    model.showPatchesChangeDialog(context);
+                                  }
                                 },
                               ),
                             ],
                           ),
-                          ...model
+                          if (model
                               .getQueriedPatches(_query)
-                              .map(
-                                (patch) => PatchItem(
-                                  name: patch.name,
-                                  simpleName: patch.getSimpleName(),
-                                  version: patch.version,
-                                  description: patch.description,
-                                  packageVersion: model.getAppVersion(),
-                                  supportedPackageVersions:
-                                      model.getSupportedVersions(patch),
-                                  isUnsupported: !isPatchSupported(patch),
-                                  isSelected: model.isSelected(patch),
-                                  onChanged: (value) =>
-                                      model.selectPatch(patch, value),
+                              .any((patch) => model.isPatchNew(patch)))
+                            Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                model.getPatchCategory(
+                                  context,
+                                  t.patchesSelectorView.newPatches,
                                 ),
-                              )
-                              .toList(),
+                                ...model.getQueriedPatches(_query).map((patch) {
+                                  if (model.isPatchNew(patch)) {
+                                    return model.getPatchItem(context, patch);
+                                  } else {
+                                    return Container();
+                                  }
+                                }),
+                                if (model.getQueriedPatches(_query).any(
+                                      (patch) =>
+                                          !model.isPatchNew(patch) &&
+                                          patch.compatiblePackages.isNotEmpty,
+                                    ))
+                                  model.getPatchCategory(
+                                    context,
+                                    t.patchesSelectorView.patches,
+                                  ),
+                              ],
+                            ),
+                          ...model.getQueriedPatches(_query).map(
+                            (patch) {
+                              if (patch.compatiblePackages.isNotEmpty &&
+                                  !model.isPatchNew(patch)) {
+                                return model.getPatchItem(context, patch);
+                              } else {
+                                return Container();
+                              }
+                            },
+                          ),
+                          if (model
+                              .getQueriedPatches(_query)
+                              .any((patch) => patch.compatiblePackages.isEmpty))
+                            Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                model.getPatchCategory(
+                                  context,
+                                  t.patchesSelectorView.universalPatches,
+                                ),
+                                ...model.getQueriedPatches(_query).map((patch) {
+                                  if (patch.compatiblePackages.isEmpty &&
+                                      !model.isPatchNew(patch)) {
+                                    return model.getPatchItem(context, patch);
+                                  } else {
+                                    return Container();
+                                  }
+                                }),
+                              ],
+                            ),
+                          const SizedBox(height: 70.0),
                         ],
                       ),
                     ),
