@@ -1,69 +1,46 @@
-import 'dart:convert';
 import 'dart:io';
-
 import 'package:collection/collection.dart';
 import 'package:dio/dio.dart';
-import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:injectable/injectable.dart';
-import 'package:revanced_manager/models/patch.dart';
+import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/services/download_manager.dart';
+import 'package:revanced_manager/services/manager_api.dart';
+import 'package:synchronized/synchronized.dart';
 
 @lazySingleton
 class GithubAPI {
-  late Dio _dio = Dio();
-
-  final _cacheOptions = CacheOptions(
-    store: MemCacheStore(),
-    maxStale: const Duration(days: 1),
-    priority: CachePriority.high,
-  );
-
-  final Map<String, String> repoAppPath = {
-    'com.google.android.youtube': 'youtube',
-    'com.google.android.apps.youtube.music': 'music',
-    'com.twitter.android': 'twitter',
-    'com.reddit.frontpage': 'reddit',
-    'com.zhiliaoapp.musically': 'tiktok',
-    'de.dwd.warnapp': 'warnwetter',
-    'com.garzotto.pflotsh.ecmwf_a': 'ecmwf',
-    'com.spotify.music': 'spotify',
-  };
+  late final Dio _dio;
+  late final ManagerAPI _managerAPI = locator<ManagerAPI>();
+  late final DownloadManager _downloadManager = locator<DownloadManager>();
+  final Map<String, Lock> _lockMap = {};
 
   Future<void> initialize(String repoUrl) async {
-    try {
-      _dio = Dio(
-        BaseOptions(
-          baseUrl: repoUrl,
-        ),
-      );
-
-      _dio.interceptors.add(DioCacheInterceptor(options: _cacheOptions));
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-    }
+    _dio = _downloadManager.initDio(repoUrl);
   }
 
   Future<void> clearAllCache() async {
-    try {
-      await _cacheOptions.store!.clean();
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
+    await _downloadManager.clearAllCache();
+  }
+
+  Future<Response> _dioGetSynchronously(String path) async {
+    // Create a new Lock for each path
+    if (!_lockMap.containsKey(path)) {
+      _lockMap[path] = Lock();
     }
+    return _lockMap[path]!.synchronized(() async {
+      return await _dio.get(path);
+    });
   }
 
   Future<Map<String, dynamic>?> getLatestRelease(
     String repoName,
   ) async {
     try {
-      final response = await _dio.get(
-        '/repos/$repoName/releases',
+      final response = await _dioGetSynchronously(
+        '/repos/$repoName/releases/latest',
       );
-      return response.data[0];
+      return response.data;
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
@@ -72,52 +49,71 @@ class GithubAPI {
     }
   }
 
-  Future<List<String>> getCommits(
-    String packageName,
-    String repoName,
-    DateTime since,
-  ) async {
-    final String path =
-        'src/main/kotlin/app/revanced/patches/${repoAppPath[packageName]}';
+  Future<String?> getChangelogs(bool isPatches) async {
+    final String repoName = isPatches
+        ? _managerAPI.getPatchesRepo()
+        : _managerAPI.defaultManagerRepo;
     try {
-      final response = await _dio.get(
-        '/repos/$repoName/commits',
-        queryParameters: {
-          'path': path,
-          'since': since.toIso8601String(),
-        },
+      final response = await _dioGetSynchronously(
+        '/repos/$repoName/releases?per_page=50',
       );
-      final List<dynamic> commits = response.data;
-      return commits
-          .map(
-            (commit) => (commit['commit']['message']).split('\n')[0] +
-                ' - ' +
-                commit['commit']['author']['name'] +
-                '\n' as String,
-          )
-          .toList();
+      final buffer = StringBuffer();
+      final String version = isPatches
+          ? _managerAPI.getLastUsedPatchesVersion()
+          : await _managerAPI.getCurrentManagerVersion();
+      int releases = 0;
+      for (final release in response.data) {
+        if (release['tag_name'] == version) {
+          if (buffer.isEmpty) {
+            buffer.writeln(release['body']);
+            releases++;
+          }
+          break;
+        }
+        if (release['prerelease']) {
+          continue;
+        }
+        buffer.writeln(release['body']);
+        releases++;
+        if (isPatches && releases == 10) {
+          break;
+        }
+      }
+      return buffer.toString();
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
+      return null;
     }
-    return [];
   }
 
-  Future<File?> getLatestReleaseFile(
+  Future<File?> getReleaseFile(
     String extension,
     String repoName,
+    String version,
+    String url,
   ) async {
     try {
-      final Map<String, dynamic>? release = await getLatestRelease(repoName);
+      if (url.isNotEmpty) {
+        return await _downloadManager.getSingleFile(
+          url,
+        );
+      }
+      final response = await _dioGetSynchronously(
+        '/repos/$repoName/releases/tags/$version',
+      );
+      final Map<String, dynamic>? release = response.data;
       if (release != null) {
         final Map<String, dynamic>? asset =
             (release['assets'] as List<dynamic>).firstWhereOrNull(
           (asset) => (asset['name'] as String).endsWith(extension),
         );
         if (asset != null) {
-          return await DefaultCacheManager().getSingleFile(
-            asset['browser_download_url'],
+          final String downloadUrl = asset['browser_download_url'];
+          _managerAPI.setPatchesDownloadURL(downloadUrl);
+          return await _downloadManager.getSingleFile(
+            downloadUrl,
           );
         }
       }
@@ -127,39 +123,5 @@ class GithubAPI {
       }
     }
     return null;
-  }
-
-  Future<List<Patch>> getPatches(String repoName) async {
-    List<Patch> patches = [];
-    try {
-      final File? f = await getLatestReleaseFile('.json', repoName);
-      if (f != null) {
-        final List<dynamic> list = jsonDecode(f.readAsStringSync());
-        patches = list.map((patch) => Patch.fromJson(patch)).toList();
-      }
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-    }
-
-    return patches;
-  }
-
-  Future<String> getLastestReleaseVersion(String repoName) async {
-    try {
-      final Map<String, dynamic>? release = await getLatestRelease(repoName);
-      if (release != null) {
-        return release['tag_name'];
-      } else {
-        return 'Unknown';
-      }
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-
-      return 'Unknown';
-    }
   }
 }
