@@ -1,57 +1,35 @@
 import 'dart:async';
-import 'dart:developer';
 import 'dart:io';
-
-import 'package:collection/collection.dart';
 import 'package:dio/dio.dart';
-import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:injectable/injectable.dart';
-import 'package:revanced_manager/models/patch.dart';
+import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/services/download_manager.dart';
+import 'package:revanced_manager/services/manager_api.dart';
+import 'package:synchronized/synchronized.dart';
 import 'package:timeago/timeago.dart';
 
 @lazySingleton
 class RevancedAPI {
-  late Dio _dio = Dio();
+  late final Dio _dio;
+  late final DownloadManager _downloadManager = locator<DownloadManager>();
 
-  final _cacheOptions = CacheOptions(
-    store: MemCacheStore(),
-    maxStale: const Duration(days: 1),
-    priority: CachePriority.high,
-  );
+  final Lock getToolsLock = Lock();
 
-  Future<void> initialize(String apiUrl) async {
-    try {
-      _dio = Dio(
-        BaseOptions(
-          baseUrl: apiUrl,
-        ),
-      );
-
-      _dio.interceptors.add(DioCacheInterceptor(options: _cacheOptions));
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-    }
+  Future<void> initialize(String repoUrl) async {
+    _dio = _downloadManager.initDio(repoUrl);
   }
 
   Future<void> clearAllCache() async {
-    try {
-      await _cacheOptions.store!.clean();
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-    }
+    await _downloadManager.clearAllCache();
   }
 
   Future<Map<String, List<dynamic>>> getContributors() async {
     final Map<String, List<dynamic>> contributors = {};
     try {
       final response = await _dio.get('/contributors');
-      final List<dynamic> repositories = response.data['repositories'];
+      final List<dynamic> repositories = response.data;
       for (final Map<String, dynamic> repo in repositories) {
         final String name = repo['name'];
         contributors[name] = repo['contributors'];
@@ -65,47 +43,31 @@ class RevancedAPI {
     return contributors;
   }
 
-  Future<List<Patch>> getPatches() async {
-    try {
-      final response = await _dio.get('/patches');
-      final List<dynamic> patches = response.data;
-      return patches.map((patch) => Patch.fromJson(patch)).toList();
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-      return List.empty();
-    }
-  }
-
   Future<Map<String, dynamic>?> _getLatestRelease(
-    String extension,
-    String repoName,
-  ) async {
-    try {
-      final response = await _dio.get('/tools');
-      final List<dynamic> tools = response.data['tools'];
-      return tools.firstWhereOrNull(
-        (t) =>
-            t['repository'] == repoName &&
-            (t['name'] as String).endsWith(extension),
-      );
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-      return null;
+    String toolName,
+  ) {
+    if (!locator<ManagerAPI>().getDownloadConsent()) {
+      return Future(() => null);
     }
+    return getToolsLock.synchronized(() async {
+      try {
+        final response = await _dio.get('/$toolName');
+        return response.data;
+      } on Exception catch (e) {
+        if (kDebugMode) {
+          print(e);
+        }
+        return null;
+      }
+    });
   }
 
   Future<String?> getLatestReleaseVersion(
-    String extension,
-    String repoName,
+    String toolName,
   ) async {
     try {
       final Map<String, dynamic>? release = await _getLatestRelease(
-        extension,
-        repoName,
+        toolName,
       );
       if (release != null) {
         return release['version'];
@@ -120,17 +82,15 @@ class RevancedAPI {
   }
 
   Future<File?> getLatestReleaseFile(
-    String extension,
-    String repoName,
+    String toolName,
   ) async {
     try {
       final Map<String, dynamic>? release = await _getLatestRelease(
-        extension,
-        repoName,
+        toolName,
       );
       if (release != null) {
-        final String url = release['browser_download_url'];
-        return await DefaultCacheManager().getSingleFile(url);
+        final String url = release['download_url'];
+        return await _downloadManager.getSingleFile(url);
       }
     } on Exception catch (e) {
       if (kDebugMode) {
@@ -141,7 +101,8 @@ class RevancedAPI {
     return null;
   }
 
-  StreamController<double> managerUpdateProgress = StreamController<double>();
+  StreamController<double> managerUpdateProgress =
+      StreamController<double>.broadcast();
 
   void updateManagerDownloadProgress(int progress) {
     managerUpdateProgress.add(progress.toDouble());
@@ -156,14 +117,10 @@ class RevancedAPI {
   }
 
   Future<File?> downloadManager() async {
-    final Map<String, dynamic>? release = await _getLatestRelease(
-      '.apk',
-      'revanced/revanced-manager',
-    );
+    final Map<String, dynamic>? release = await _getLatestRelease('manager');
     File? outputFile;
-    await for (final result in DefaultCacheManager().getFileStream(
-      release!['browser_download_url'] as String,
-      withProgress: true,
+    await for (final result in _downloadManager.getFileStream(
+      release!['download_url'] as String,
     )) {
       if (result is DownloadProgress) {
         final totalSize = result.totalSize ?? 10000000;
@@ -171,6 +128,7 @@ class RevancedAPI {
 
         updateManagerDownloadProgress(progress);
       } else if (result is FileInfo) {
+        disposeManagerUpdateProgress();
         // The download is complete; convert the FileInfo object to a File object
         outputFile = File(result.file.path);
       }
@@ -179,17 +137,15 @@ class RevancedAPI {
   }
 
   Future<String?> getLatestReleaseTime(
-    String extension,
-    String repoName,
+    String toolName,
   ) async {
     try {
       final Map<String, dynamic>? release = await _getLatestRelease(
-        extension,
-        repoName,
+        toolName,
       );
       if (release != null) {
         final DateTime timestamp =
-            DateTime.parse(release['timestamp'] as String);
+            DateTime.parse(release['created_at'] as String);
         return format(timestamp, locale: 'en_short');
       }
     } on Exception catch (e) {
