@@ -2,8 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:root/root.dart';
 
 class RootAPI {
-  final String _managerDirPath = '/data/local/tmp/revanced-manager';
-  final String _postFsDataDirPath = '/data/adb/post-fs-data.d';
+  final String _revancedDirPath = '/data/adb/revanced';
   final String _serviceDDirPath = '/data/adb/service.d';
 
   Future<bool> isRooted() async {
@@ -40,96 +39,84 @@ class RootAPI {
     seLinux,
     String filePath,
   ) async {
-    if (permissions.isNotEmpty) {
+    try {
+      final StringBuffer commands = StringBuffer();
+      if (permissions.isNotEmpty) {
+        commands.writeln('chmod $permissions $filePath');
+      }
+
+      if (ownerGroup.isNotEmpty) {
+        commands.writeln('chown $ownerGroup $filePath');
+      }
+
+      if (seLinux.isNotEmpty) {
+        commands.writeln('chcon $seLinux $filePath');
+      }
+
       await Root.exec(
-        cmd: 'chmod $permissions "$filePath"',
+        cmd: commands.toString(),
       );
-    }
-    if (ownerGroup.isNotEmpty) {
-      await Root.exec(
-        cmd: 'chown $ownerGroup "$filePath"',
-      );
-    }
-    if (seLinux.isNotEmpty) {
-      await Root.exec(
-        cmd: 'chcon $seLinux "$filePath"',
-      );
+    } on Exception catch (e) {
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
   Future<bool> isAppInstalled(String packageName) async {
     if (packageName.isNotEmpty) {
-      String? res = await Root.exec(
-        cmd: 'ls "$_managerDirPath/$packageName"',
-      );
-      if (res != null && res.isNotEmpty) {
-        res = await Root.exec(
-          cmd: 'ls "$_serviceDDirPath/$packageName.sh"',
-        );
-        return res != null && res.isNotEmpty;
-      }
+      return fileExists('$_serviceDDirPath/$packageName.sh');
     }
     return false;
   }
 
   Future<List<String>> getInstalledApps() async {
+    final List<String> apps = List.empty(growable: true);
     try {
-      final String? res = await Root.exec(
-        cmd: 'ls "$_managerDirPath"',
-      );
+      final String? res = await Root.exec(cmd: 'ls $_revancedDirPath');
       if (res != null) {
-        final List<String> apps = res.split('\n');
-        apps.removeWhere((pack) => pack.isEmpty);
-        return apps.map((pack) => pack.trim()).toList();
+        final List<String> list = res.split('\n');
+        list.removeWhere((pack) => pack.isEmpty);
+        apps.addAll(list.map((pack) => pack.trim()).toList());
       }
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
-      return List.empty();
     }
-    return List.empty();
+    return apps;
   }
 
-  Future<void> deleteApp(String packageName, String originalFilePath) async {
+  Future<void> uninstall(String packageName) async {
     await Root.exec(
-      cmd: 'am force-stop "$packageName"',
-    );
-    await Root.exec(
-      cmd: 'su -mm -c "umount -l $originalFilePath"',
-    );
-    await Root.exec(
-      cmd: 'rm -rf "$_managerDirPath/$packageName"',
-    );
-    await Root.exec(
-      cmd: 'rm -rf "$_serviceDDirPath/$packageName.sh"',
-    );
-    await Root.exec(
-      cmd: 'rm -rf "$_postFsDataDirPath/$packageName.sh"',
+      cmd: '''
+          grep $packageName /proc/mounts | while read -r line; do echo \$line | cut -d " " -f 2 | sed "s/apk.*/apk/" | xargs -r umount -l; done
+          rm -rf $_revancedDirPath/$packageName $_serviceDDirPath/$packageName.sh
+          ''',
     );
   }
 
-  Future<bool> installApp(
+  Future<void> removeOrphanedFiles() async {
+    await Root.exec(
+      cmd: 'find $_revancedDirPath -type f -name original.apk -delete',
+    );
+  }
+
+  Future<bool> install(
     String packageName,
     String originalFilePath,
     String patchedFilePath,
   ) async {
     try {
-      await deleteApp(packageName, originalFilePath);
-      await Root.exec(
-        cmd: 'mkdir -p "$_managerDirPath/$packageName"',
-      );
       await setPermissions(
         '0755',
         'shell:shell',
         '',
-        '$_managerDirPath/$packageName',
+        '$_revancedDirPath/$packageName',
       );
-      await saveOriginalFilePath(packageName, originalFilePath);
+      await installPatchedApk(packageName, patchedFilePath);
       await installServiceDScript(packageName);
-      await installPostFsDataScript(packageName);
-      await installApk(packageName, patchedFilePath);
-      await mountApk(packageName, originalFilePath);
+      await runMountScript(packageName);
       return true;
     } on Exception catch (e) {
       if (kDebugMode) {
@@ -140,33 +127,49 @@ class RootAPI {
   }
 
   Future<void> installServiceDScript(String packageName) async {
-    final String content = '#!/system/bin/sh\n'
-        'while [ "\$(getprop sys.boot_completed | tr -d \'"\'"\'\\\\r\'"\'"\')" != "1" ]; do sleep 3; done\n'
-        'base_path=$_managerDirPath/$packageName/base.apk\n'
-        'stock_path=\$(pm path $packageName | grep base | sed \'"\'"\'s/package://g\'"\'"\')\n'
-        r'[ ! -z $stock_path ] && mount -o bind $base_path $stock_path';
+    await Root.exec(
+      cmd: 'mkdir -p $_serviceDDirPath',
+    );
+    final String mountScript = '''
+    #!/system/bin/sh
+    # Mount using Magisk mirror, if available.
+    MAGISKTMP="\$( magisk --path )" || MAGISKTMP=/sbin
+    MIRROR="\$MAGISKTMP/.magisk/mirror"
+    if [ ! -f \$MIRROR ]; then
+        MIRROR=""
+    fi
+
+    until [ "\$(getprop sys.boot_completed)" = 1 ]; do sleep 3; done
+    until [ -d "/sdcard/Android" ]; do sleep 1; done
+    
+    # Unmount any existing installation to prevent multiple unnecessary mounts.
+    grep $packageName /proc/mounts | while read -r line; do echo \$line | cut -d " " -f 2 | sed "s/apk.*/apk/" | xargs -r umount -l; done
+
+    base_path=$_revancedDirPath/$packageName/base.apk
+    stock_path=\$(pm path $packageName | grep base | sed "s/package://g" )
+
+    chcon u:object_r:apk_data_file:s0  \$base_path
+    mount -o bind \$MIRROR\$base_path \$stock_path
+
+    # Kill the app to force it to restart the mounted APK in case it is already running
+    am force-stop $packageName
+    '''
+        .trimMultilineString();
     final String scriptFilePath = '$_serviceDDirPath/$packageName.sh';
     await Root.exec(
-      cmd: 'echo \'$content\' > "$scriptFilePath"',
+      cmd: 'echo \'$mountScript\' > "$scriptFilePath"',
     );
     await setPermissions('0744', '', '', scriptFilePath);
   }
 
-  Future<void> installPostFsDataScript(String packageName) async {
-    final String content = '#!/system/bin/sh\n'
-        'stock_path=\$(pm path $packageName | grep base | sed \'"\'"\'s/package://g\'"\'"\')\n'
-        r'[ ! -z $stock_path ] && umount -l $stock_path';
-    final String scriptFilePath = '$_postFsDataDirPath/$packageName.sh';
+  Future<void> installPatchedApk(
+      String packageName, String patchedFilePath,) async {
+    final String newPatchedFilePath = '$_revancedDirPath/$packageName/base.apk';
     await Root.exec(
-      cmd: 'echo \'$content\' > "$scriptFilePath"',
-    );
-    await setPermissions('0744', '', '', scriptFilePath);
-  }
-
-  Future<void> installApk(String packageName, String patchedFilePath) async {
-    final String newPatchedFilePath = '$_managerDirPath/$packageName/base.apk';
-    await Root.exec(
-      cmd: 'cp "$patchedFilePath" "$newPatchedFilePath"',
+      cmd: '''
+      mkdir -p $_revancedDirPath/$packageName
+      cp "$patchedFilePath" $newPatchedFilePath
+      ''',
     );
     await setPermissions(
       '0644',
@@ -176,66 +179,30 @@ class RootAPI {
     );
   }
 
-  Future<void> mountApk(String packageName, String originalFilePath) async {
-    final String newPatchedFilePath = '$_managerDirPath/$packageName/base.apk';
-    await Root.exec(
-      cmd: 'am force-stop "$packageName"',
-    );
-    await Root.exec(
-      cmd: 'su -mm -c "umount -l $originalFilePath"',
-    );
-    await Root.exec(
-      cmd: 'su -mm -c "mount -o bind $newPatchedFilePath $originalFilePath"',
-    );
-  }
-
-  Future<bool> isMounted(String packageName) async {
-    final String? res = await Root.exec(
-      cmd: 'cat /proc/mounts | grep $packageName',
-    );
-    return res != null && res.isNotEmpty;
-  }
-
-  Future<String> getOriginalFilePath(
+  Future<void> runMountScript(
     String packageName,
-    String originalFilePath,
   ) async {
-    final bool isInstalled = await isAppInstalled(packageName);
-    if (isInstalled && await isMounted(packageName)) {
-      originalFilePath = '$_managerDirPath/$packageName/original.apk';
-      await setPermissions(
-        '0644',
-        'shell:shell',
-        'u:object_r:apk_data_file:s0',
-        originalFilePath,
+    await Root.exec(cmd: '.$_serviceDDirPath/$packageName.sh');
+  }
+
+  Future<bool> fileExists(String path) async {
+    try {
+      final String? res = await Root.exec(
+        cmd: 'ls $path',
       );
+      return res != null && res.isNotEmpty;
+    } on Exception catch (e) {
+      if (kDebugMode) {
+        print(e);
+      }
+      return false;
     }
-    return originalFilePath;
   }
+}
 
-  Future<void> saveOriginalFilePath(
-    String packageName,
-    String originalFilePath,
-  ) async {
-    final String originalRootPath =
-        '$_managerDirPath/$packageName/original.apk';
-    await Root.exec(
-      cmd: 'mkdir -p "$_managerDirPath/$packageName"',
-    );
-    await setPermissions(
-      '0755',
-      'shell:shell',
-      '',
-      '$_managerDirPath/$packageName',
-    );
-    await Root.exec(
-      cmd: 'cp "$originalFilePath" "$originalRootPath"',
-    );
-    await setPermissions(
-      '0644',
-      'shell:shell',
-      'u:object_r:apk_data_file:s0',
-      originalFilePath,
-    );
-  }
+// Remove leading spaces manually until
+// https://github.com/dart-lang/language/issues/559 is closed
+extension StringExtension on String {
+  String trimMultilineString() =>
+      split('\n').map((line) => line.trim()).join('\n').trim();
 }
