@@ -2,16 +2,16 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:app_installer/app_installer.dart';
-import 'package:cross_connectivity/cross_connectivity.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:injectable/injectable.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:revanced_manager/app/app.locator.dart';
 import 'package:revanced_manager/app/app.router.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/models/patched_application.dart';
 import 'package:revanced_manager/services/github_api.dart';
 import 'package:revanced_manager/services/manager_api.dart';
@@ -20,9 +20,8 @@ import 'package:revanced_manager/services/revanced_api.dart';
 import 'package:revanced_manager/services/toast.dart';
 import 'package:revanced_manager/ui/views/navigation/navigation_viewmodel.dart';
 import 'package:revanced_manager/ui/views/patcher/patcher_viewmodel.dart';
-import 'package:revanced_manager/ui/widgets/homeView/update_confirmation_dialog.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_material_button.dart';
-import 'package:revanced_manager/utils/about_info.dart';
+import 'package:revanced_manager/ui/widgets/homeView/update_confirmation_sheet.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_checkbox_list_tile.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -35,29 +34,48 @@ class HomeViewModel extends BaseViewModel {
   final RevancedAPI _revancedAPI = locator<RevancedAPI>();
   final Toast _toast = locator<Toast>();
   final flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
-  DateTime? _lastUpdate;
   bool showUpdatableApps = false;
+  PatchedApplication? lastPatchedApp;
+  bool releaseBuild = false;
   List<PatchedApplication> patchedInstalledApps = [];
-  List<PatchedApplication> patchedUpdatableApps = [];
-  String _managerVersion = '';
+  String _currentManagerVersion = '';
+  String _currentPatchesVersion = '';
+  String? latestManagerVersion;
+  String? latestPatchesVersion;
+  File? downloadedApk;
 
   Future<void> initialize(BuildContext context) async {
-    _managerVersion = await AboutInfo.getInfo().then(
-      (value) => value.keys.contains('version') ? value['version']! : '',
-    );
-    _managerVersion = await _managerAPI.getCurrentManagerVersion();
+    _managerAPI.reAssessPatchedApps().then((_) => getPatchedApps());
+    _currentManagerVersion = await _managerAPI.getCurrentManagerVersion();
+    if (!_managerAPI.getDownloadConsent()) {
+      await showDownloadConsent(context);
+      await forceRefresh(context);
+      return;
+    }
+    _currentPatchesVersion = await _managerAPI.getCurrentPatchesVersion();
+    if (_managerAPI.showUpdateDialog() && await hasManagerUpdates()) {
+      showUpdateDialog(context, false);
+    }
+    if (!_managerAPI.isPatchesAutoUpdate() &&
+        _managerAPI.showUpdateDialog() &&
+        await hasPatchesUpdates()) {
+      showUpdateDialog(context, true);
+    }
+
+    await _patcherAPI.initialize();
+
     await flutterLocalNotificationsPlugin.initialize(
       const InitializationSettings(
         android: AndroidInitializationSettings('ic_notification'),
       ),
       onDidReceiveNotificationResponse: (response) async {
         if (response.id == 0) {
-          _toast.showBottom('homeView.installingMessage');
+          _toast.showBottom(t.homeView.installingMessage);
           final File? managerApk = await _managerAPI.downloadManager();
           if (managerApk != null) {
-            await AppInstaller.installApk(managerApk.path);
+            await _patcherAPI.installApk(context, managerApk.path);
           } else {
-            _toast.showBottom('homeView.errorDownloadMessage');
+            _toast.showBottom(t.homeView.errorDownloadMessage);
           }
         }
       },
@@ -65,30 +83,31 @@ class HomeViewModel extends BaseViewModel {
     flutterLocalNotificationsPlugin
         .resolvePlatformSpecificImplementation<
             AndroidFlutterLocalNotificationsPlugin>()
-        ?.requestPermission();
-    final bool isConnected = await Connectivity().checkConnection();
+        ?.requestNotificationsPermission();
+    final bool isConnected = !(await Connectivity().checkConnectivity())
+        .contains(ConnectivityResult.none);
     if (!isConnected) {
-      _toast.showBottom('homeView.noConnection');
+      _toast.showBottom(t.homeView.noConnection);
     }
+
     final NotificationAppLaunchDetails? notificationAppLaunchDetails =
         await flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails();
     if (notificationAppLaunchDetails?.didNotificationLaunchApp ?? false) {
-      _toast.showBottom('homeView.installingMessage');
+      _toast.showBottom(t.homeView.installingMessage);
       final File? managerApk = await _managerAPI.downloadManager();
       if (managerApk != null) {
-        await AppInstaller.installApk(managerApk.path);
+        await _patcherAPI.installApk(context, managerApk.path);
       } else {
-        _toast.showBottom('homeView.errorDownloadMessage');
+        _toast.showBottom(t.homeView.errorDownloadMessage);
       }
     }
-    _getPatchedApps();
-    _managerAPI.reAssessSavedApps().then((_) => _getPatchedApps());
   }
 
-  void navigateToAppInfo(PatchedApplication app) {
+  void navigateToAppInfo(PatchedApplication app, bool isLastPatchedApp) {
     _navigationService.navigateTo(
       Routes.appInfoView,
-      arguments: AppInfoViewArguments(app: app),
+      arguments:
+          AppInfoViewArguments(app: app, isLastPatchedApp: isLastPatchedApp),
     );
   }
 
@@ -105,39 +124,37 @@ class HomeViewModel extends BaseViewModel {
     locator<NavigationViewModel>().setIndex(1);
   }
 
-  void _getPatchedApps() {
+  void getPatchedApps() {
+    lastPatchedApp = _managerAPI.getLastPatchedApp();
     patchedInstalledApps = _managerAPI.getPatchedApps().toList();
-    patchedUpdatableApps = _managerAPI
-        .getPatchedApps()
-        .where((app) => app.hasUpdates == true)
-        .toList();
     notifyListeners();
   }
 
+  bool isLastPatchedAppEnabled() {
+    return _managerAPI.isLastPatchedAppEnabled();
+  }
+
   Future<bool> hasManagerUpdates() async {
-    final String? latestVersion = await _managerAPI.getLatestManagerVersion();
-    String currentVersion = await _managerAPI.getCurrentManagerVersion();
-
-    // add v to current version
-    if (!currentVersion.startsWith('v')) {
-      currentVersion = 'v$currentVersion';
+    if (!_managerAPI.releaseBuild) {
+      return false;
     }
+    latestManagerVersion =
+        await _managerAPI.getLatestManagerVersion() ?? _currentManagerVersion;
 
-    if (latestVersion != currentVersion) {
+    if (latestManagerVersion != _currentManagerVersion) {
       return true;
     }
     return false;
   }
 
   Future<bool> hasPatchesUpdates() async {
-    final String? latestVersion = await _managerAPI.getLatestPatchesVersion();
-    final String? currentVersion = await _managerAPI.getCurrentPatchesVersion();
-    if (latestVersion != null) {
+    latestPatchesVersion = await _managerAPI.getLatestPatchesVersion();
+    if (latestPatchesVersion != null) {
       try {
         final int latestVersionInt =
-            int.parse(latestVersion.replaceAll(RegExp('[^0-9]'), ''));
+            int.parse(latestPatchesVersion!.replaceAll(RegExp('[^0-9]'), ''));
         final int currentVersionInt =
-            int.parse(currentVersion!.replaceAll(RegExp('[^0-9]'), ''));
+            int.parse(_currentPatchesVersion.replaceAll(RegExp('[^0-9]'), ''));
         return latestVersionInt > currentVersionInt;
       } on Exception catch (e) {
         if (kDebugMode) {
@@ -165,75 +182,243 @@ class HomeViewModel extends BaseViewModel {
     }
   }
 
-  Future<void> updateManager(BuildContext context) async {
-    try {
-      _toast.showBottom('homeView.downloadingMessage');
-      showDialog(
-        context: context,
-        builder: (context) => SimpleDialog(
-          contentPadding: const EdgeInsets.all(16.0),
-          title: I18nText(
-            'homeView.downloadingMessage',
-            child: Text(
-              '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w500,
-                color: Theme.of(context).colorScheme.secondary,
-              ),
-            ),
-          ),
-          children: [
-            Column(
-              children: [
-                Row(
-                  children: [
-                    Icon(
-                      Icons.new_releases_outlined,
+  Future<void> showDownloadConsent(BuildContext context) async {
+    final ValueNotifier<bool> autoUpdate = ValueNotifier(true);
+    await showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) => PopScope(
+        canPop: false,
+        child: AlertDialog(
+          title: Text(t.homeView.downloadConsentDialogTitle),
+          content: ValueListenableBuilder(
+            valueListenable: autoUpdate,
+            builder: (context, value, child) {
+              return Column(
+                mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    t.homeView.downloadConsentDialogText,
+                    style: TextStyle(
+                      fontSize: 16,
+                      fontWeight: FontWeight.w500,
                       color: Theme.of(context).colorScheme.secondary,
                     ),
-                    const SizedBox(width: 8.0),
-                    Text(
-                      'v$_managerVersion',
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.symmetric(vertical: 10),
+                    child: Text(
+                      t.homeView.downloadConsentDialogText2(
+                        url: _managerAPI.defaultApiUrl.split('/')[2],
+                      ),
                       style: TextStyle(
-                        fontSize: 18,
+                        fontSize: 16,
                         fontWeight: FontWeight.w500,
-                        color: Theme.of(context).colorScheme.secondary,
+                        color: Theme.of(context).colorScheme.error,
                       ),
                     ),
-                  ],
-                ),
-                const SizedBox(height: 16.0),
-                StreamBuilder<double>(
-                  initialData: 0.0,
-                  stream: _revancedAPI.managerUpdateProgress.stream,
-                  builder: (context, snapshot) {
-                    return LinearProgressIndicator(
-                      value: snapshot.data! * 0.01,
-                      valueColor: AlwaysStoppedAnimation<Color>(
-                        Theme.of(context).colorScheme.secondary,
-                      ),
-                    );
-                  },
-                ),
-                const SizedBox(height: 16.0),
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: CustomMaterialButton(
-                    label: I18nText('cancelButton'),
-                    onPressed: () {
-                      _revancedAPI.disposeManagerUpdateProgress();
-                      Navigator.of(context).pop();
-                    },
                   ),
-                ),
-              ],
+                ],
+              );
+            },
+          ),
+          actions: [
+            TextButton(
+              onPressed: () async {
+                _managerAPI.setDownloadConsent(false);
+                SystemNavigator.pop();
+              },
+              child: Text(t.quitButton),
+            ),
+            FilledButton(
+              onPressed: () async {
+                _managerAPI.setDownloadConsent(true);
+                _managerAPI.setPatchesAutoUpdate(autoUpdate.value);
+                Navigator.of(context).pop();
+              },
+              child: Text(t.okButton),
             ),
           ],
+        ),
+      ),
+    );
+  }
+
+  void showUpdateDialog(BuildContext context, bool isPatches) {
+    final ValueNotifier<bool> noShow =
+        ValueNotifier(!_managerAPI.showUpdateDialog());
+    showDialog(
+      context: context,
+      builder: (innerContext) => AlertDialog(
+        title: Text(t.homeView.updateDialogTitle),
+        content: ValueListenableBuilder(
+          valueListenable: noShow,
+          builder: (context, value, child) {
+            return Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  t.homeView.updateDialogText(
+                    file: isPatches ? 'ReVanced Patches' : 'ReVanced Manager',
+                    version: isPatches
+                        ? _currentPatchesVersion
+                        : _currentManagerVersion,
+                  ),
+                  style: TextStyle(
+                    fontSize: 16,
+                    fontWeight: FontWeight.w500,
+                    color: Theme.of(context).colorScheme.secondary,
+                  ),
+                ),
+                const SizedBox(height: 10),
+                HapticCheckboxListTile(
+                  value: value,
+                  contentPadding: EdgeInsets.zero,
+                  title: Text(t.noShowAgain),
+                  subtitle: Text(t.homeView.changeLaterSubtitle),
+                  onChanged: (selected) {
+                    noShow.value = selected!;
+                  },
+                ),
+              ],
+            );
+          },
+        ),
+        actions: [
+          TextButton(
+            onPressed: () async {
+              _managerAPI.setShowUpdateDialog(!noShow.value);
+              Navigator.pop(innerContext);
+            },
+            child: Text(t.dismissButton), // Decide later
+          ),
+          FilledButton(
+            onPressed: () async {
+              _managerAPI.setShowUpdateDialog(!noShow.value);
+              Navigator.pop(innerContext);
+              await showUpdateConfirmationDialog(context, isPatches);
+            },
+            child: Text(t.showUpdateButton),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> updatePatches(BuildContext context) async {
+    _toast.showBottom(t.homeView.downloadingMessage);
+    final String patchesVersion =
+        await _managerAPI.getLatestPatchesVersion() ?? '0.0.0';
+    if (patchesVersion != '0.0.0') {
+      await _managerAPI.setCurrentPatchesVersion(patchesVersion);
+      _toast.showBottom(t.homeView.downloadedMessage);
+      forceRefresh(context);
+    } else {
+      _toast.showBottom(t.homeView.errorDownloadMessage);
+    }
+  }
+
+  Future<void> updateManager(BuildContext context) async {
+    final ValueNotifier<bool> downloaded = ValueNotifier(false);
+    try {
+      _toast.showBottom(t.homeView.downloadingMessage);
+      showDialog(
+        context: context,
+        builder: (context) => ValueListenableBuilder(
+          valueListenable: downloaded,
+          builder: (context, value, child) {
+            return AlertDialog(
+              title: Text(
+                !value
+                    ? t.homeView.downloadingMessage
+                    : t.homeView.downloadedMessage,
+              ),
+              content: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  if (!value)
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        StreamBuilder<double>(
+                          initialData: 0.0,
+                          stream: _revancedAPI.managerUpdateProgress.stream,
+                          builder: (context, snapshot) {
+                            return LinearProgressIndicator(
+                              value: snapshot.data! * 0.01,
+                              valueColor: AlwaysStoppedAnimation<Color>(
+                                Theme.of(context).colorScheme.secondary,
+                              ),
+                            );
+                          },
+                        ),
+                        const SizedBox(height: 16.0),
+                        Align(
+                          alignment: Alignment.centerRight,
+                          child: FilledButton(
+                            onPressed: () {
+                              _revancedAPI.disposeManagerUpdateProgress();
+                              Navigator.of(context).pop();
+                            },
+                            child: Text(t.cancelButton),
+                          ),
+                        ),
+                      ],
+                    ),
+                  if (value)
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          t.homeView.installUpdate,
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.w500,
+                            color: Theme.of(context).colorScheme.secondary,
+                          ),
+                        ),
+                        const SizedBox(height: 16.0),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.end,
+                          children: [
+                            Align(
+                              alignment: Alignment.centerRight,
+                              child: TextButton(
+                                onPressed: () {
+                                  Navigator.of(context).pop();
+                                },
+                                child: Text(t.cancelButton),
+                              ),
+                            ),
+                            const SizedBox(width: 8.0),
+                            Align(
+                              alignment: Alignment.centerRight,
+                              child: FilledButton(
+                                onPressed: () async {
+                                  await _patcherAPI.installApk(
+                                    context,
+                                    downloadedApk!.path,
+                                  );
+                                },
+                                child: Text(t.updateButton),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                ],
+              ),
+            );
+          },
         ),
       );
       final File? managerApk = await downloadManager();
       if (managerApk != null) {
+        downloaded.value = true;
+        downloadedApk = managerApk;
         // await flutterLocalNotificationsPlugin.zonedSchedule(
         //   0,
         //   FlutterI18n.translate(
@@ -258,42 +443,44 @@ class HomeViewModel extends BaseViewModel {
         //   uiLocalNotificationDateInterpretation:
         //       UILocalNotificationDateInterpretation.absoluteTime,
         // );
-        _toast.showBottom('homeView.installingMessage');
-        await AppInstaller.installApk(managerApk.path);
+        _toast.showBottom(t.homeView.installingMessage);
+        await _patcherAPI.installApk(context, managerApk.path);
       } else {
-        _toast.showBottom('homeView.errorDownloadMessage');
+        _toast.showBottom(t.homeView.errorDownloadMessage);
       }
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
-      _toast.showBottom('homeView.errorInstallMessage');
+      _toast.showBottom(t.homeView.errorInstallMessage);
     }
-  }
-
-  void updatesAreDisabled() {
-    _toast.showBottom('homeView.updatesDisabled');
   }
 
   Future<void> showUpdateConfirmationDialog(
     BuildContext parentContext,
-  ) {
+    bool isPatches, [
+    bool changelog = false,
+  ]) {
     return showModalBottomSheet(
       context: parentContext,
+      useSafeArea: true,
       isScrollControlled: true,
       shape: const RoundedRectangleBorder(
         borderRadius: BorderRadius.vertical(top: Radius.circular(24.0)),
       ),
-      builder: (context) => const UpdateConfirmationDialog(),
+      builder: (context) => UpdateConfirmationSheet(
+        isPatches: isPatches,
+        changelog: changelog,
+      ),
     );
   }
 
-  Future<Map<String, dynamic>?> getLatestManagerRelease() {
-    return _githubAPI.getLatestRelease(_managerAPI.defaultManagerRepo);
+  Future<String?> getChangelogs(bool isPatches) {
+    return _githubAPI.getChangelogs(isPatches);
   }
 
-  Future<String?> getLatestPatcherReleaseTime() {
-    return _managerAPI.getLatestPatcherReleaseTime();
+  Future<String?> getLatestPatchesReleaseTime() {
+    return _managerAPI.getLatestPatchesReleaseTime();
   }
 
   Future<String?> getLatestManagerReleaseTime() {
@@ -301,12 +488,8 @@ class HomeViewModel extends BaseViewModel {
   }
 
   Future<void> forceRefresh(BuildContext context) async {
-    await Future.delayed(const Duration(seconds: 1));
-    if (_lastUpdate == null ||
-        _lastUpdate!.difference(DateTime.now()).inSeconds > 2) {
-      _managerAPI.clearAllData();
-    }
-    _toast.showBottom('homeView.refreshSuccess');
-    initialize(context);
+    await _managerAPI.clearAllData();
+    await initialize(context);
+    _toast.showBottom(t.homeView.refreshSuccess);
   }
 }
