@@ -1,75 +1,71 @@
 // ignore_for_file: use_build_context_synchronously
 
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/services/manager_api.dart';
-import 'package:revanced_manager/services/toast.dart';
-import 'package:revanced_manager/ui/widgets/settingsView/custom_text_field.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_tile_dialog.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_material_button.dart';
 import 'package:stacked/stacked.dart';
 
 class SManageApiUrl extends BaseViewModel {
   final ManagerAPI _managerAPI = locator<ManagerAPI>();
-  final Toast _toast = locator<Toast>();
 
   final TextEditingController _apiUrlController = TextEditingController();
 
   Future<void> showApiUrlDialog(BuildContext context) async {
-    final String apiUrl = _managerAPI.getApiUrl();
-    _apiUrlController.text = apiUrl.replaceAll('https://', '');
+    final apiUrl = _managerAPI.getApiUrl();
+
+    _apiUrlController.text = apiUrl;
     return showDialog(
       context: context,
       builder: (context) => AlertDialog(
         title: Row(
           children: <Widget>[
-            I18nText('settingsView.apiURLLabel'),
-            const Spacer(),
+            Expanded(
+              child: Text(t.settingsView.apiURLLabel),
+            ),
             IconButton(
               icon: const Icon(Icons.manage_history_outlined),
               onPressed: () => showApiUrlResetDialog(context),
               color: Theme.of(context).colorScheme.secondary,
-            )
+            ),
           ],
         ),
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
         content: SingleChildScrollView(
           child: Column(
             children: <Widget>[
-              CustomTextField(
-                leadingIcon: Icon(
-                  Icons.api_outlined,
-                  color: Theme.of(context).colorScheme.secondary,
-                ),
-                inputController: _apiUrlController,
-                label: I18nText('settingsView.selectApiURL'),
-                hint: apiUrl,
+              TextField(
+                controller: _apiUrlController,
+                autocorrect: false,
                 onChanged: (value) => notifyListeners(),
+                decoration: InputDecoration(
+                  icon: Icon(
+                    Icons.api_outlined,
+                    color: Theme.of(context).colorScheme.onSurfaceVariant,
+                  ),
+                  border: const OutlineInputBorder(),
+                  labelText: t.settingsView.selectApiURL,
+                  hintText: apiUrl,
+                ),
               ),
             ],
           ),
         ),
         actions: <Widget>[
-          CustomMaterialButton(
-            isFilled: false,
-            label: I18nText('cancelButton'),
+          TextButton(
             onPressed: () {
               _apiUrlController.clear();
               Navigator.of(context).pop();
             },
+            child: Text(t.cancelButton),
           ),
-          CustomMaterialButton(
-            label: I18nText('okButton'),
+          FilledButton(
             onPressed: () {
-              String apiUrl = _apiUrlController.text;
-              if (!apiUrl.startsWith('https')) {
-                apiUrl = 'https://$apiUrl';
-              }
-              _managerAPI.setApiUrl(apiUrl);
+              _managerAPI.setApiUrl(_apiUrlController.text);
               Navigator.of(context).pop();
             },
-          )
+            child: Text(t.okButton),
+          ),
         ],
       ),
     );
@@ -79,23 +75,22 @@ class SManageApiUrl extends BaseViewModel {
     return showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: I18nText('settingsView.sourcesResetDialogTitle'),
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
-        content: I18nText('settingsView.apiURLResetDialogText'),
+        title: Text(t.settingsView.sourcesResetDialogTitle),
+        content: Text(t.settingsView.apiURLResetDialogText),
         actions: <Widget>[
-          CustomMaterialButton(
-            isFilled: false,
-            label: I18nText('noButton'),
+          TextButton(
             onPressed: () => Navigator.of(context).pop(),
+            child: Text(t.noButton),
           ),
-          CustomMaterialButton(
-            label: I18nText('yesButton'),
+          FilledButton(
             onPressed: () {
-              _managerAPI.setApiUrl('');
-              _toast.showBottom('settingsView.restartAppForChanges');
-              Navigator.of(context).pop();
+              _managerAPI.resetApiUrl();
+              Navigator.of(context)
+                ..pop()
+                ..pop();
             },
-          )
+            child: Text(t.yesButton),
+          ),
         ],
       ),
     );
@@ -111,8 +106,8 @@ class SManageApiUrlUI extends StatelessWidget {
   Widget build(BuildContext context) {
     return SettingsTileDialog(
       padding: const EdgeInsets.symmetric(horizontal: 20.0),
-      title: 'settingsView.apiURLLabel',
-      subtitle: 'settingsView.apiURLHint',
+      title: t.settingsView.apiURLLabel,
+      subtitle: t.settingsView.apiURLHint,
       onTap: () => sManageApiUrl.showApiUrlDialog(context),
     );
   }
