@@ -1,135 +1,97 @@
 // ignore_for_file: use_build_context_synchronously
 
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/services/manager_api.dart';
 import 'package:revanced_manager/services/toast.dart';
-import 'package:revanced_manager/ui/widgets/settingsView/custom_text_field.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_tile_dialog.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_material_button.dart';
 import 'package:stacked/stacked.dart';
 
 class SManageSources extends BaseViewModel {
   final ManagerAPI _managerAPI = locator<ManagerAPI>();
   final Toast _toast = locator<Toast>();
 
-  final TextEditingController _hostSourceController = TextEditingController();
   final TextEditingController _orgPatSourceController = TextEditingController();
   final TextEditingController _patSourceController = TextEditingController();
-  final TextEditingController _orgIntSourceController = TextEditingController();
-  final TextEditingController _intSourceController = TextEditingController();
 
   Future<void> showSourcesDialog(BuildContext context) async {
-    final String hostRepository = _managerAPI.getRepoUrl();
     final String patchesRepo = _managerAPI.getPatchesRepo();
-    final String integrationsRepo = _managerAPI.getIntegrationsRepo();
-    _hostSourceController.text = hostRepository;
     _orgPatSourceController.text = patchesRepo.split('/')[0];
     _patSourceController.text = patchesRepo.split('/')[1];
-    _orgIntSourceController.text = integrationsRepo.split('/')[0];
-    _intSourceController.text = integrationsRepo.split('/')[1];
     return showDialog(
       context: context,
       builder: (context) => AlertDialog(
+        scrollable: true,
         title: Row(
           children: <Widget>[
-            I18nText('settingsView.sourcesLabel'),
-            const Spacer(),
+            Expanded(
+              child: Text(t.settingsView.sourcesLabel),
+            ),
             IconButton(
               icon: const Icon(Icons.manage_history_outlined),
               onPressed: () => showResetConfirmationDialog(context),
               color: Theme.of(context).colorScheme.secondary,
-            )
+            ),
           ],
         ),
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
-        content: SingleChildScrollView(
-          child: Column(
-            children: <Widget>[
-              CustomTextField(
-                leadingIcon: const Icon(
+        content: Column(
+          children: <Widget>[
+            TextField(
+              controller: _orgPatSourceController,
+              autocorrect: false,
+              onChanged: (value) => notifyListeners(),
+              decoration: InputDecoration(
+                icon: Icon(
+                  Icons.extension_outlined,
+                  color: Theme.of(context).colorScheme.onSurfaceVariant,
+                ),
+                border: const OutlineInputBorder(),
+                labelText: t.settingsView.orgPatchesLabel,
+                hintText: patchesRepo.split('/')[0],
+              ),
+            ),
+            const SizedBox(height: 8),
+            // Patches repository's name
+            TextField(
+              controller: _patSourceController,
+              autocorrect: false,
+              onChanged: (value) => notifyListeners(),
+              decoration: InputDecoration(
+                icon: const Icon(
                   Icons.extension_outlined,
                   color: Colors.transparent,
                 ),
-                inputController: _hostSourceController,
-                label: I18nText('settingsView.hostRepositoryLabel'),
-                hint: hostRepository,
-                onChanged: (value) => notifyListeners(),
+                border: const OutlineInputBorder(),
+                labelText: t.settingsView.sourcesPatchesLabel,
+                hintText: patchesRepo.split('/')[1],
               ),
-              const SizedBox(height: 20),
-              CustomTextField(
-                leadingIcon: Icon(
-                  Icons.extension_outlined,
-                  color: Theme.of(context).colorScheme.secondary,
-                ),
-                inputController: _orgPatSourceController,
-                label: I18nText('settingsView.orgPatchesLabel'),
-                hint: patchesRepo.split('/')[0],
-                onChanged: (value) => notifyListeners(),
-              ),
-              const SizedBox(height: 8),
-              CustomTextField(
-                leadingIcon: const Icon(
-                  Icons.extension_outlined,
-                  color: Colors.transparent,
-                ),
-                inputController: _patSourceController,
-                label: I18nText('settingsView.sourcesPatchesLabel'),
-                hint: patchesRepo.split('/')[1],
-                onChanged: (value) => notifyListeners(),
-              ),
-              const SizedBox(height: 20),
-              CustomTextField(
-                leadingIcon: Icon(
-                  Icons.merge_outlined,
-                  color: Theme.of(context).colorScheme.secondary,
-                ),
-                inputController: _orgIntSourceController,
-                label: I18nText('settingsView.orgIntegrationsLabel'),
-                hint: integrationsRepo.split('/')[0],
-                onChanged: (value) => notifyListeners(),
-              ),
-              const SizedBox(height: 8),
-              CustomTextField(
-                leadingIcon: const Icon(
-                  Icons.merge_outlined,
-                  color: Colors.transparent,
-                ),
-                inputController: _intSourceController,
-                label: I18nText('settingsView.sourcesIntegrationsLabel'),
-                hint: integrationsRepo.split('/')[1],
-                onChanged: (value) => notifyListeners(),
-              ),
-            ],
-          ),
+            ),
+            const SizedBox(height: 20),
+            Text(t.settingsView.sourcesUpdateNote),
+          ],
         ),
         actions: <Widget>[
-          CustomMaterialButton(
-            isFilled: false,
-            label: I18nText('cancelButton'),
+          TextButton(
             onPressed: () {
               _orgPatSourceController.clear();
               _patSourceController.clear();
-              _orgIntSourceController.clear();
-              _intSourceController.clear();
               Navigator.of(context).pop();
             },
+            child: Text(t.cancelButton),
           ),
-          CustomMaterialButton(
-            label: I18nText('okButton'),
+          FilledButton(
             onPressed: () {
-              _managerAPI.setRepoUrl(_hostSourceController.text.trim());
               _managerAPI.setPatchesRepo(
                 '${_orgPatSourceController.text.trim()}/${_patSourceController.text.trim()}',
               );
-              _managerAPI.setIntegrationsRepo(
-                '${_orgIntSourceController.text.trim()}/${_intSourceController.text.trim()}',
-              );
-              _toast.showBottom('settingsView.restartAppForChanges');
+              _managerAPI.setCurrentPatchesVersion('0.0.0');
+              _managerAPI.setLastUsedPatchesVersion();
+              _toast.showBottom(t.settingsView.restartAppForChanges);
               Navigator.of(context).pop();
             },
-          )
+            child: Text(t.okButton),
+          ),
         ],
       ),
     );
@@ -139,27 +101,24 @@ class SManageSources extends BaseViewModel {
     return showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: I18nText('settingsView.sourcesResetDialogTitle'),
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
-        content: I18nText('settingsView.sourcesResetDialogText'),
+        title: Text(t.settingsView.sourcesResetDialogTitle),
+        content: Text(t.settingsView.sourcesResetDialogText),
         actions: <Widget>[
-          CustomMaterialButton(
-            isFilled: false,
-            label: I18nText('noButton'),
+          TextButton(
             onPressed: () => Navigator.of(context).pop(),
+            child: Text(t.noButton),
           ),
-          CustomMaterialButton(
-            label: I18nText('yesButton'),
+          FilledButton(
             onPressed: () {
-              _managerAPI.setRepoUrl('');
               _managerAPI.setPatchesRepo('');
-              _managerAPI.setIntegrationsRepo('');
-              _toast.showBottom('settingsView.restartAppForChanges');
+              _managerAPI.setCurrentPatchesVersion('0.0.0');
+              _toast.showBottom(t.settingsView.restartAppForChanges);
               Navigator.of(context)
                 ..pop()
                 ..pop();
             },
-          )
+            child: Text(t.yesButton),
+          ),
         ],
       ),
     );
@@ -175,8 +134,8 @@ class SManageSourcesUI extends StatelessWidget {
   Widget build(BuildContext context) {
     return SettingsTileDialog(
       padding: const EdgeInsets.symmetric(horizontal: 20.0),
-      title: 'settingsView.sourcesLabel',
-      subtitle: 'settingsView.sourcesLabelHint',
+      title: t.settingsView.sourcesLabel,
+      subtitle: t.settingsView.sourcesLabelHint,
       onTap: () => sManageSources.showSourcesDialog(context),
     );
   }
