@@ -3,114 +3,176 @@
 import 'package:dynamic_themes/dynamic_themes.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_i18n/widgets/I18nText.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/services/manager_api.dart';
-import 'package:revanced_manager/ui/views/settings/settings_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_section.dart';
-import 'package:stacked/stacked.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_radio_list_tile.dart';
+import 'package:revanced_manager/ui/widgets/shared/haptics/haptic_switch_list_tile.dart';
 
-final _settingViewModel = SettingsViewModel();
+class SUpdateThemeUI extends StatefulWidget {
+  const SUpdateThemeUI({super.key});
 
-// ignore: constant_identifier_names
-const int ANDROID_12_SDK_VERSION = 31;
-
-class SUpdateTheme extends BaseViewModel {
-  final ManagerAPI _managerAPI = locator<ManagerAPI>();
-
-  bool getDynamicThemeStatus() {
-    return _managerAPI.getUseDynamicTheme();
-  }
-
-  Future<void> setUseDynamicTheme(BuildContext context, bool value) async {
-    await _managerAPI.setUseDynamicTheme(value);
-    final int currentTheme = DynamicTheme.of(context)!.themeId;
-    if (currentTheme.isEven) {
-      await DynamicTheme.of(context)!.setTheme(value ? 2 : 0);
-    } else {
-      await DynamicTheme.of(context)!.setTheme(value ? 3 : 1);
-    }
-    notifyListeners();
-  }
-
-  bool getDarkThemeStatus() {
-    return _managerAPI.getUseDarkTheme();
-  }
-
-  Future<void> setUseDarkTheme(BuildContext context, bool value) async {
-    await _managerAPI.setUseDarkTheme(value);
-    final int currentTheme = DynamicTheme.of(context)!.themeId;
-    if (currentTheme < 2) {
-      await DynamicTheme.of(context)!.setTheme(value ? 1 : 0);
-    } else {
-      await DynamicTheme.of(context)!.setTheme(value ? 3 : 2);
-    }
-    SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(
-        systemNavigationBarIconBrightness:
-            value ? Brightness.light : Brightness.dark,
-      ),
-    );
-    notifyListeners();
-  }
+  @override
+  State<SUpdateThemeUI> createState() => _SUpdateThemeUIState();
 }
 
-class SUpdateThemeUI extends StatelessWidget {
-  const SUpdateThemeUI({super.key});
+class _SUpdateThemeUIState extends State<SUpdateThemeUI> {
+  final ManagerAPI managerAPI = locator<ManagerAPI>();
 
   @override
   Widget build(BuildContext context) {
     return SettingsSection(
-      title: 'settingsView.appearanceSectionTitle',
+      title: t.settingsView.appearanceSectionTitle,
       children: <Widget>[
-        SwitchListTile(
+        ListTile(
           contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
-          title: I18nText(
-            'settingsView.darkThemeLabel',
-            child: const Text(
-              '',
-              style: TextStyle(
+          title: Text(
+            t.settingsView.themeModeLabel,
+            style: const TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.w500,
+            ),
+          ),
+          trailing: FilledButton(
+            onPressed: () => {showThemeDialog(context)},
+            child: getThemeModeName(),
+          ),
+          onTap: () => {showThemeDialog(context)},
+        ),
+        if (managerAPI.isDynamicThemeAvailable)
+          HapticSwitchListTile(
+            contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
+            title: Text(
+              t.settingsView.dynamicThemeLabel,
+              style: const TextStyle(
                 fontSize: 20,
                 fontWeight: FontWeight.w500,
               ),
             ),
-          ),
-          subtitle: I18nText('settingsView.darkThemeHint'),
-          value: SUpdateTheme().getDarkThemeStatus(),
-          onChanged: (value) => SUpdateTheme().setUseDarkTheme(
-            context,
-            value,
-          ),
-        ),
-        FutureBuilder<int>(
-          future: _settingViewModel.getSdkVersion(),
-          builder: (context, snapshot) => Visibility(
-            visible:
-                snapshot.hasData && snapshot.data! >= ANDROID_12_SDK_VERSION,
-            child: SwitchListTile(
-              contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
-              title: I18nText(
-                'settingsView.dynamicThemeLabel',
-                child: const Text(
-                  '',
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
+            subtitle: Text(t.settingsView.dynamicThemeHint),
+            value: getDynamicThemeStatus(),
+            onChanged: (value) => {
+              setUseDynamicTheme(
+                context,
+                value,
               ),
-              subtitle: I18nText('settingsView.dynamicThemeHint'),
-              value: _settingViewModel.sUpdateTheme.getDynamicThemeStatus(),
-              onChanged: (value) => {
-                _settingViewModel.sUpdateTheme.setUseDynamicTheme(
-                  context,
-                  value,
-                ),
-              },
-            ),
+            },
+          ),
+      ],
+    );
+  }
+
+  bool getDynamicThemeStatus() {
+    return managerAPI.getUseDynamicTheme();
+  }
+
+  Future<void> setUseDynamicTheme(BuildContext context, bool value) async {
+    await managerAPI.setUseDynamicTheme(value);
+    final int currentTheme = (DynamicTheme.of(context)!.themeId ~/ 2) * 2;
+    await DynamicTheme.of(context)!.setTheme(currentTheme + (value ? 1 : 0));
+    setState(() {});
+  }
+
+  int getThemeMode() {
+    return managerAPI.getThemeMode();
+  }
+
+  Future<void> setThemeMode(BuildContext context, int value) async {
+    await managerAPI.setThemeMode(value);
+    final bool isDynamicTheme = DynamicTheme.of(context)!.themeId.isEven;
+    await DynamicTheme.of(context)!
+        .setTheme(value * 2 + (isDynamicTheme ? 0 : 1));
+    final bool isLight = value != 2 &&
+        (value == 1 ||
+            DynamicTheme.of(context)!.theme.brightness == Brightness.light);
+    SystemChrome.setSystemUIOverlayStyle(
+      SystemUiOverlayStyle(
+        systemNavigationBarIconBrightness:
+            isLight ? Brightness.dark : Brightness.light,
+      ),
+    );
+    setState(() {});
+  }
+
+  Text getThemeModeName() {
+    switch (getThemeMode()) {
+      case 0:
+        return Text(t.settingsView.systemThemeLabel);
+      case 1:
+        return Text(t.settingsView.lightThemeLabel);
+      case 2:
+        return Text(t.settingsView.darkThemeLabel);
+      default:
+        return Text(t.settingsView.systemThemeLabel);
+    }
+  }
+
+  Future<void> showThemeDialog(BuildContext context) async {
+    final ValueNotifier<int> newTheme = ValueNotifier(getThemeMode());
+
+    return showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: Text(t.settingsView.themeModeLabel),
+        icon: const Icon(Icons.palette),
+        contentPadding: const EdgeInsets.symmetric(vertical: 16),
+        content: SingleChildScrollView(
+          child: ValueListenableBuilder(
+            valueListenable: newTheme,
+            builder: (context, value, child) {
+              return Column(
+                mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  HapticRadioListTile(
+                    title: Text(t.settingsView.systemThemeLabel),
+                    contentPadding: const EdgeInsets.symmetric(horizontal: 16),
+                    value: 0,
+                    groupValue: value,
+                    onChanged: (value) {
+                      newTheme.value = value!;
+                    },
+                  ),
+                  HapticRadioListTile(
+                    title: Text(t.settingsView.lightThemeLabel),
+                    contentPadding: const EdgeInsets.symmetric(horizontal: 16),
+                    value: 1,
+                    groupValue: value,
+                    onChanged: (value) {
+                      newTheme.value = value!;
+                    },
+                  ),
+                  HapticRadioListTile(
+                    title: Text(t.settingsView.darkThemeLabel),
+                    contentPadding: const EdgeInsets.symmetric(horizontal: 16),
+                    value: 2,
+                    groupValue: value,
+                    onChanged: (value) {
+                      newTheme.value = value!;
+                    },
+                  ),
+                ],
+              );
+            },
           ),
         ),
-      ],
+        actions: <Widget>[
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: Text(t.cancelButton),
+          ),
+          FilledButton(
+            onPressed: () {
+              setThemeMode(context, newTheme.value);
+              Navigator.of(context).pop();
+            },
+            child: Text(t.okButton),
+          ),
+        ],
+      ),
     );
   }
 }
