@@ -1,23 +1,29 @@
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_card.dart';
 
 class InstalledAppItem extends StatefulWidget {
   const InstalledAppItem({
-    Key? key,
+    super.key,
     required this.name,
     required this.pkgName,
     required this.icon,
     required this.patchesCount,
     required this.suggestedVersion,
+    required this.installedVersion,
     this.onTap,
-  }) : super(key: key);
+    this.onLinkTap,
+  });
+
   final String name;
   final String pkgName;
   final Uint8List icon;
   final int patchesCount;
   final String suggestedVersion;
+  final String installedVersion;
   final Function()? onTap;
+  final Function()? onLinkTap;
 
   @override
   State<InstalledAppItem> createState() => _InstalledAppItemState();
@@ -48,31 +54,77 @@ class _InstalledAppItemState extends State<InstalledAppItem> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
-                  Text(
-                    widget.name,
-                    maxLines: 2,
-                    overflow: TextOverflow.visible,
-                    style: const TextStyle(
-                      fontSize: 16,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(widget.pkgName),
-                  Row(
+                  Wrap(
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    spacing: 4,
                     children: [
                       Text(
-                        widget.suggestedVersion.isEmpty
-                            ? 'All versions'
-                            : widget.suggestedVersion,
+                        widget.name,
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
+                        style: const TextStyle(
+                          fontSize: 16,
+                        ),
                       ),
-                      const SizedBox(width: 4),
+                      Text(
+                        widget.installedVersion,
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
+                        style: const TextStyle(
+                          fontSize: 16,
+                        ),
+                      ),
                       Text(
                         widget.patchesCount == 1
                             ? '• ${widget.patchesCount} patch'
                             : '• ${widget.patchesCount} patches',
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
                         style: TextStyle(
+                          fontSize: 16,
                           color: Theme.of(context).colorScheme.secondary,
+                        ),
+                      ),
+                    ],
+                  ),
+                  Text(
+                    widget.pkgName,
+                  ),
+                  const SizedBox(height: 4),
+                  Wrap(
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    children: [
+                      Material(
+                        color: Theme.of(context).colorScheme.secondaryContainer,
+                        borderRadius:
+                            const BorderRadius.all(Radius.circular(8)),
+                        child: InkWell(
+                          onTap: widget.onLinkTap,
+                          borderRadius:
+                              const BorderRadius.all(Radius.circular(8)),
+                          child: Container(
+                            padding: const EdgeInsets.fromLTRB(8, 4, 8, 4),
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                Text(
+                                  t.suggested(
+                                    version: widget.suggestedVersion.isEmpty
+                                        ? t.appSelectorCard.anyVersion
+                                        : 'v${widget.suggestedVersion}',
+                                  ),
+                                ),
+                                const SizedBox(width: 4),
+                                Icon(
+                                  Icons.search,
+                                  size: 16,
+                                  color: Theme.of(context)
+                                      .colorScheme
+                                      .onSecondaryContainer,
+                                ),
+                              ],
+                            ),
+                          ),
                         ),
                       ),
                     ],
