@@ -1,18 +1,22 @@
 import 'package:flutter/material.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_card.dart';
 
 class NotInstalledAppItem extends StatefulWidget {
   const NotInstalledAppItem({
-    Key? key,
+    super.key,
     required this.name,
     required this.patchesCount,
     required this.suggestedVersion,
     this.onTap,
-  }) : super(key: key);
+    this.onLinkTap,
+  });
+
   final String name;
   final int patchesCount;
   final String suggestedVersion;
   final Function()? onTap;
+  final Function()? onLinkTap;
 
   @override
   State<NotInstalledAppItem> createState() => _NotInstalledAppItem();
@@ -29,15 +33,15 @@ class _NotInstalledAppItem extends State<NotInstalledAppItem> {
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
             Container(
+              width: 48,
               height: 48,
-              padding: const EdgeInsets.symmetric(vertical: 4.0),
               alignment: Alignment.center,
               child: const CircleAvatar(
                 backgroundColor: Colors.transparent,
                 child: Icon(
                   Icons.square_rounded,
                   color: Colors.grey,
-                  size: 44,
+                  size: 48,
                 ),
               ),
             ),
@@ -46,30 +50,65 @@ class _NotInstalledAppItem extends State<NotInstalledAppItem> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
-                  Text(
-                    widget.name,
-                    style: const TextStyle(
-                      fontSize: 16,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  const Text('App not installed.'),
-                  const SizedBox(height: 4),
-                  Row(
+                  Wrap(
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    spacing: 4,
                     children: [
                       Text(
-                        widget.suggestedVersion.isEmpty
-                            ? 'All versions'
-                            : widget.suggestedVersion,
+                        widget.name,
+                        style: const TextStyle(
+                          fontSize: 16,
+                        ),
                       ),
-                      const SizedBox(width: 4),
                       Text(
                         widget.patchesCount == 1
                             ? '• ${widget.patchesCount} patch'
                             : '• ${widget.patchesCount} patches',
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
                         style: TextStyle(
+                          fontSize: 16,
                           color: Theme.of(context).colorScheme.secondary,
+                        ),
+                      ),
+                    ],
+                  ),
+                  const SizedBox(height: 4),
+                  Wrap(
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    children: [
+                      Material(
+                        color:
+                            Theme.of(context).colorScheme.secondaryContainer,
+                        borderRadius:
+                            const BorderRadius.all(Radius.circular(8)),
+                        child: InkWell(
+                          onTap: widget.onLinkTap,
+                          borderRadius:
+                              const BorderRadius.all(Radius.circular(8)),
+                          child: Container(
+                            padding: const EdgeInsets.fromLTRB(8, 4, 8, 4),
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                Text(
+                                  t.suggested(
+                                    version: widget.suggestedVersion.isEmpty
+                                        ? t.appSelectorCard.anyVersion
+                                        : 'v${widget.suggestedVersion}',
+                                  ),
+                                ),
+                                const SizedBox(width: 4),
+                                Icon(
+                                  Icons.search,
+                                  size: 16,
+                                  color: Theme.of(context)
+                                      .colorScheme
+                                      .onSecondaryContainer,
+                                ),
+                              ],
+                            ),
+                          ),
                         ),
                       ),
                     ],
