@@ -1,16 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/file.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:revanced_manager/services/download_manager.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_card.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class ContributorsCard extends StatefulWidget {
   const ContributorsCard({
-    Key? key,
+    super.key,
     required this.title,
     required this.contributors,
-  }) : super(key: key);
+  });
   final String title;
   final List<dynamic> contributors;
 
@@ -26,14 +25,11 @@ class _ContributorsCardState extends State<ContributorsCard> {
       children: <Widget>[
         Padding(
           padding: const EdgeInsets.only(bottom: 8.0),
-          child: I18nText(
+          child: Text(
             widget.title,
-            child: const Text(
-              '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w500,
-              ),
+            style: const TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.w500,
             ),
           ),
         ),
@@ -53,12 +49,12 @@ class _ContributorsCardState extends State<ContributorsCard> {
               child: GestureDetector(
                 onTap: () => launchUrl(
                   Uri.parse(
-                    widget.contributors[index]['html_url'],
+                    widget.contributors[index]['url'],
                   ),
                   mode: LaunchMode.externalApplication,
                 ),
                 child: FutureBuilder<File?>(
-                  future: DefaultCacheManager().getSingleFile(
+                  future: DownloadManager().getSingleFile(
                     widget.contributors[index]['avatar_url'],
                   ),
                   builder: (context, snapshot) => snapshot.hasData
