@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/ui/views/home/home_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_card.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_material_button.dart';
 
 class LatestCommitCard extends StatefulWidget {
   const LatestCommitCard({
-    Key? key,
-    required this.onPressedManager,
-    required this.onPressedPatches,
-  }) : super(key: key);
-  final Function() onPressedManager;
-  final Function() onPressedPatches;
+    super.key,
+    required this.model,
+    required this.parentContext,
+  });
+  final HomeViewModel model;
+  final BuildContext parentContext;
 
   @override
   State<LatestCommitCard> createState() => _LatestCommitCardState();
@@ -30,42 +29,41 @@ class _LatestCommitCardState extends State<LatestCommitCard> {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Row(
-                    children: const <Widget>[
-                      Text('ReVanced Manager'),
-                    ],
-                  ),
-                  const SizedBox(height: 4),
-                  Row(
-                    children: <Widget>[
-                      FutureBuilder<String?>(
-                        future: model.getLatestManagerReleaseTime(),
-                        builder: (context, snapshot) =>
-                            snapshot.hasData && snapshot.data!.isNotEmpty
-                                ? I18nText(
-                                    'latestCommitCard.timeagoLabel',
-                                    translationParams: {'time': snapshot.data!},
-                                  )
-                                : I18nText('latestCommitCard.loadingLabel'),
-                      ),
-                    ],
-                  ),
-                ],
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    const Text('ReVanced Manager'),
+                    const SizedBox(height: 4),
+                    Row(
+                      children: <Widget>[
+                        FutureBuilder<String?>(
+                          future: model.getLatestManagerReleaseTime(),
+                          builder: (context, snapshot) =>
+                              snapshot.hasData && snapshot.data!.isNotEmpty
+                                  ? Text(
+                                      t.latestCommitCard
+                                          .timeagoLabel(time: snapshot.data!),
+                                    )
+                                  : Text(t.latestCommitCard.loadingLabel),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
               ),
               FutureBuilder<bool>(
                 future: model.hasManagerUpdates(),
                 initialData: false,
-                builder: (context, snapshot) => Opacity(
-                  opacity: snapshot.hasData && snapshot.data! ? 1.0 : 0.25,
-                  child: CustomMaterialButton(
-                    label: I18nText('updateButton'),
-                    onPressed: snapshot.hasData && snapshot.data!
-                        ? widget.onPressedManager
-                        : () => {},
+                builder: (context, snapshot) => FilledButton(
+                  onPressed: () => widget.model.showUpdateConfirmationDialog(
+                    widget.parentContext,
+                    false,
+                    !snapshot.data!,
                   ),
+                  child: (snapshot.hasData && !snapshot.data!)
+                      ? Text(t.showChangelogButton)
+                      : Text(t.showUpdateButton),
                 ),
               ),
             ],
@@ -74,52 +72,45 @@ class _LatestCommitCardState extends State<LatestCommitCard> {
 
         const SizedBox(height: 16),
 
-        // ReVanced Patches
+        // Patches
         CustomCard(
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Row(
-                    children: const <Widget>[
-                      Text('ReVanced Patches'),
-                    ],
-                  ),
-                  const SizedBox(height: 4),
-                  Row(
-                    children: <Widget>[
-                      FutureBuilder<String?>(
-                        future: model.getLatestPatcherReleaseTime(),
-                        builder: (context, snapshot) => Text(
-                          snapshot.hasData && snapshot.data!.isNotEmpty
-                              ? FlutterI18n.translate(
-                                  context,
-                                  'latestCommitCard.timeagoLabel',
-                                  translationParams: {'time': snapshot.data!},
-                                )
-                              : FlutterI18n.translate(
-                                  context,
-                                  'latestCommitCard.loadingLabel',
-                                ),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    const Text('ReVanced Patches'),
+                    const SizedBox(height: 4),
+                    Row(
+                      children: <Widget>[
+                        FutureBuilder<String?>(
+                          future: model.getLatestPatchesReleaseTime(),
+                          builder: (context, snapshot) => Text(
+                            snapshot.hasData && snapshot.data!.isNotEmpty
+                                ? t.latestCommitCard
+                                    .timeagoLabel(time: snapshot.data!)
+                                : t.latestCommitCard.loadingLabel,
+                          ),
                         ),
-                      ),
-                    ],
-                  ),
-                ],
+                      ],
+                    ),
+                  ],
+                ),
               ),
               FutureBuilder<bool>(
-                future: locator<HomeViewModel>().hasPatchesUpdates(),
+                future: model.hasPatchesUpdates(),
                 initialData: false,
-                builder: (context, snapshot) => Opacity(
-                  opacity: snapshot.hasData && snapshot.data! ? 1.0 : 0.25,
-                  child: CustomMaterialButton(
-                    label: I18nText('updateButton'),
-                    onPressed: snapshot.hasData && snapshot.data!
-                        ? widget.onPressedPatches
-                        : () => {},
+                builder: (context, snapshot) => FilledButton(
+                  onPressed: () => widget.model.showUpdateConfirmationDialog(
+                    widget.parentContext,
+                    true,
+                    !snapshot.data!,
                   ),
+                  child: (snapshot.hasData && !snapshot.data!)
+                      ? Text(t.showChangelogButton)
+                      : Text(t.showUpdateButton),
                 ),
               ),
             ],
