@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:revanced_manager/app/app.locator.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
 import 'package:revanced_manager/models/patch.dart';
 import 'package:revanced_manager/ui/views/patcher/patcher_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/shared/custom_card.dart';
 
 class PatchSelectorCard extends StatelessWidget {
   const PatchSelectorCard({
-    Key? key,
+    super.key,
     required this.onPressed,
-  }) : super(key: key);
+  });
   final Function() onPressed;
 
   @override
@@ -21,16 +21,13 @@ class PatchSelectorCard extends StatelessWidget {
         children: <Widget>[
           Row(
             children: <Widget>[
-              I18nText(
+              Text(
                 locator<PatcherViewModel>().selectedPatches.isEmpty
-                    ? 'patchSelectorCard.widgetTitle'
-                    : 'patchSelectorCard.widgetTitleSelected',
-                child: const Text(
-                  '',
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                  ),
+                    ? t.patchSelectorCard.widgetTitle
+                    : t.patchSelectorCard.widgetTitleSelected,
+                style: const TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.w500,
                 ),
               ),
               Text(
@@ -46,10 +43,10 @@ class PatchSelectorCard extends StatelessWidget {
           ),
           const SizedBox(height: 4),
           if (locator<PatcherViewModel>().selectedApp == null)
-            I18nText('patchSelectorCard.widgetSubtitle')
+            Text(t.patchSelectorCard.widgetSubtitle)
           else
             locator<PatcherViewModel>().selectedPatches.isEmpty
-                ? I18nText('patchSelectorCard.widgetEmptySubtitle')
+                ? Text(t.patchSelectorCard.widgetEmptySubtitle)
                 : Text(_getPatchesSelection()),
         ],
       ),
@@ -58,8 +55,11 @@ class PatchSelectorCard extends StatelessWidget {
 
   String _getPatchesSelection() {
     String text = '';
-    for (final Patch p in locator<PatcherViewModel>().selectedPatches) {
-      text += '\u2022  ${p.getSimpleName()} (v${p.version})\n';
+    final List<Patch> selectedPatches =
+        locator<PatcherViewModel>().selectedPatches;
+    selectedPatches.sort((a, b) => a.name.compareTo(b.name));
+    for (final Patch p in selectedPatches) {
+      text += '•  ${p.getSimpleName()}\n';
     }
     return text.substring(0, text.length - 1);
   }
