@@ -1,16 +1,15 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/widgets/I18nText.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_manage_api_url.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_manage_sources.dart';
-import 'package:revanced_manager/ui/views/settings/settings_viewmodel.dart';
-import 'package:revanced_manager/ui/widgets/settingsView/settings_experimental_patches.dart';
-import 'package:revanced_manager/ui/widgets/settingsView/settings_experimental_universal_patches.dart';
+import 'package:revanced_manager/gen/strings.g.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_auto_update_patches.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_enable_patches_selection.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_last_patched_app.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_require_suggested_app_version.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_section.dart';
-import 'package:revanced_manager/ui/widgets/shared/custom_material_button.dart';
-
-final _settingsViewModel = SettingsViewModel();
+import 'package:revanced_manager/ui/widgets/settingsView/settings_show_update_dialog.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_universal_patches.dart';
+import 'package:revanced_manager/ui/widgets/settingsView/settings_version_compatibility_check.dart';
 
 class SAdvancedSection extends StatelessWidget {
   const SAdvancedSection({super.key});
@@ -18,86 +17,16 @@ class SAdvancedSection extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SettingsSection(
-      title: 'settingsView.advancedSectionTitle',
-      children: <Widget>[
-        SManageApiUrlUI(),
-        SManageSourcesUI(),
-        // SManageKeystorePasswordUI(),
-        SExperimentalUniversalPatches(),
-        SExperimentalPatches(),
-        ListTile(
-          contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
-          title: I18nText(
-            'settingsView.deleteKeystoreLabel',
-            child: const Text(
-              '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-          subtitle: I18nText('settingsView.deleteKeystoreHint'),
-          onTap: () => _showDeleteKeystoreDialog(context),
-        ),
-        ListTile(
-          contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
-          title: I18nText(
-            'settingsView.deleteTempDirLabel',
-            child: const Text(
-              '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-          subtitle: I18nText('settingsView.deleteTempDirHint'),
-          onTap: () => _settingsViewModel.deleteTempDir(),
-        ),
-        ListTile(
-          contentPadding: const EdgeInsets.symmetric(horizontal: 20.0),
-          title: I18nText(
-            'settingsView.deleteLogsLabel',
-            child: const Text(
-              '',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-          subtitle: I18nText('settingsView.deleteLogsHint'),
-          onTap: () => _settingsViewModel.deleteLogs(),
-        ),
+      title: t.settingsView.advancedSectionTitle,
+      children: const <Widget>[
+        SAutoUpdatePatches(),
+        SShowUpdateDialog(),
+        SEnablePatchesSelection(),
+        SRequireSuggestedAppVersion(),
+        SVersionCompatibilityCheck(),
+        SUniversalPatches(),
+        SLastPatchedApp(),
       ],
-    );
-  }
-
-  Future<void> _showDeleteKeystoreDialog(context) {
-    return showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: I18nText('warning'),
-        backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
-        content: I18nText(
-          'settingsView.deleteKeystoreDialogText',
-        ),
-        actions: <Widget>[
-          CustomMaterialButton(
-            isFilled: false,
-            label: I18nText('noButton'),
-            onPressed: () => Navigator.of(context).pop(),
-          ),
-          CustomMaterialButton(
-            label: I18nText('yesButton'),
-            onPressed: () => {
-              Navigator.of(context).pop(),
-              _settingsViewModel.deleteKeystore()
-            },
-          )
-        ],
-      ),
     );
   }
 }
